% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_probes.R
\name{add_probes}
\alias{add_probes}
\title{Add probes to (RT)-PCR assays}
\usage{
add_probes(x, y, tm_difference = c(0, 20))
}
\arguments{
\item{x}{Assays to add probes to (an object of class 'rprimer_assay').}

\item{y}{Candidate probes (an object of class 'rprimer_oligo').}

\item{tm_difference}{Acceptable Tm difference between primers and probe.
The minimum allowed value is -20 and the maximum allowed value is 20.
The default are \code{c(0, 20)}.}
}
\value{
Assays with probes. A tibble (a data frame) of
class 'rprimer_assay'.

##### Describe more here #####
}
\description{
\code{add_probes} adds probes to (RT)-PCR assays.
}
\details{
The Tm-difference is calculated by subtracting the
Tm of the probe with the average Tm of the
primer pair. Thus, a negative Tm-difference
means that the Tm of the probe is lower than the average Tm of the
primer pair.

Note that the Tm-difference is calculated from the majority oligos, and
may thus be misleading for degenerate (iupac) oligos.
}
\examples{
add_probes(
example_rprimer_assay,
example_rprimer_oligo,
tm_difference = c(-2, 10)
)

}
\seealso{
get_assays
}
