% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignmentProperties.R
\name{asIUPAC}
\alias{asIUPAC}
\title{Convert DNA nucleotides into the corresponding IUPAC base}
\usage{
asIUPAC(x)
}
\arguments{
\item{x}{A character vector of length one, containing DNA
nucleotides (valid bases are A, C, G, T, - and .). Each base must
be separated by a comma (,), e.g. 'A,C,G'.
Characters other than A, C, G, T, and - will be ignored. However, when
the input only consist of invalid bases,
or if the bases are not separated by ',',
\code{asIUPAC} will return NA.}
}
\value{
The corresponding IUPAC base.
}
\description{
\code{asIUPAC} takes several DNA nucleotides as input,
and returns the degenerate base in IUPAC format.
}
\examples{
asIUPAC("A,C")
asIUPAC("R")
asIUPAC("TG") ## Will return NA since the bases are not separated by comma

}
\keyword{internal}
