% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{check_match}
\alias{check_match}
\alias{check_match.rprimer_oligo}
\alias{check_match.rprimer_assay}
\title{Check if oligos or assays matches their targets (generic)}
\usage{
check_match(x, y)

\method{check_match}{rprimer_oligo}(x, y)

\method{check_match}{rprimer_assay}(x, y)
}
\arguments{
\item{x}{An object of class
'rprimer_oligo' or 'rprimer_assay'.}

\item{y}{The intended target. An alignment of DNA sequences
(an object of class 'rprimer_alignment').}
}
\value{
A tibble (
a data frame) of class 'rprimer_oligo' or 'rprimer_assay',
with columns describing the proportion of perfectly matching
sequences for each oligo or assay, and a column named
"match_report", which contains a matrix with information about
which sequences the oligo/assay matches perfectly to.
}
\description{
\code{check_match} checks if oligos or assays matches with their
intended target sequences.
}
\section{Methods (by class)}{
\itemize{
\item \code{rprimer_oligo}: Check match of an object of class
'rprimer_oligo'.

\item \code{rprimer_assay}: Check match of an object of class
'rprimer_assay'.
}}

\examples{
check_match(example_rprimer_oligo, example_rprimer_alignment)
check_match(example_rprimer_assay, example_rprimer_alignment)

}
