% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToDNAStringSet.R
\name{convertToDNAStringSet}
\alias{convertToDNAStringSet}
\alias{convertToDNAStringSet,RprimerOligo-method}
\alias{convertToDNAStringSet,RprimerAssay-method}
\title{Convert to DNA string set (generic)}
\usage{
convertToDNAStringSet(x, revAsRc = TRUE)

\S4method{convertToDNAStringSet}{RprimerOligo}(x, revAsRc = TRUE)

\S4method{convertToDNAStringSet}{RprimerAssay}(x, revAsRc = TRUE)
}
\arguments{
\item{x}{An \code{RprimerProfile}, \code{RprimerOligo} or \code{RprimerAssay} object.}

\item{revAsRc}{If reverse primers/negative sense probes should be written as reverse
complements. Defaults to \code{TRUE}.}
}
\value{
A Biostrings::DNAStringSet
}
\description{
\code{convertToDNAStringSet()} can be used for converting all sequence
variants of all oligos or assays to a \code{Biostrings::DNAStringSet} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{RprimerOligo}: 

\item \code{RprimerAssay}: 
}}

\examples{
data("exampleRprimerOligo")
convertToDNAStringSet(exampleRprimerOligo[1:2, ])

data("exampleRprimerAssay")
convertToDNAStringSet(exampleRprimerAssay[1:2, ], revAsRc = FALSE)
}
\references{
H. Pages, P. Aboyoun, R. Gentleman and S. DebRoy (2020). Biostrings:
Efficient manipulation of biological strings. R package version
2.57.2.
}
