% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_degeneracy.R
\name{count_degeneracy}
\alias{count_degeneracy}
\title{Count the degeneracy of a DNA sequence}
\usage{
count_degeneracy(x)
}
\arguments{
\item{x}{a DNA sequence (a character vector of length one, e.g. 'cttgg').}
}
\value{
The number of unique sequences of x (an integer).
}
\description{
\code{count_degenerates} counts the number of unique sequences of
a DNA sequence with degenerate bases.
}
\details{
Valid bases for \code{x} are 'a', 'c', 'g', 't', 'r', 'y', 'm',
'k', 's', 'w', n', 'h', 'd', 'v', 'b' and '-'.
}
\examples{
count_degeneracy("cttnra")

}
