% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_degenerates.R
\name{count_degenerates}
\alias{count_degenerates}
\title{Count the number of degenerate bases in a DNA sequence}
\usage{
count_degenerates(x)
}
\arguments{
\item{x}{a DNA sequence (a character vector of length one, e.g. 'cttgg').}
}
\value{
the number of degenerate bases in \code{x} (an integer).
}
\description{
\code{count_degenerates} returns the number of degenerate bases in a DNA
sequence.
}
\details{
Valid bases for \code{x} are 'a', 'c', 'g', 't', 'r', 'y', 'm',
'k', 's', 'w', n', 'h', 'd', 'v', 'b' and '-'.
}
\examples{
count_degenerates("cttnra")

}
