% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDegeneracy.R
\name{countDegeneracy}
\alias{countDegeneracy}
\title{Count the degeneracy of a DNA sequence}
\usage{
countDegeneracy(x)
}
\arguments{
\item{x}{A DNA sequence (a character vector of length one).
Valid bases are ACGTRYSWKMBDHVN-.}
}
\value{
The number of unique sequences of x (an integer).
}
\description{
\code{countDegeneracy} returns the number of unique variants of
a DNA sequence with degenerate bases.
}
\keyword{internal}
