% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDegenerates.R
\name{countDegenerates}
\alias{countDegenerates}
\title{Count the number of degenerate bases in a DNA sequence}
\usage{
countDegenerates(x)
}
\arguments{
\item{x}{A DNA sequence (a character vector of length one).
Valid bases are ACGTRYSWKMBDHVN-.}
}
\value{
The number of degenerate bases in \code{x} (an integer).
}
\description{
\code{countDegenerates} returns the number of degenerate bases in a DNA
sequence.
}
\examples{
count_degenerates("CTTRNA")

}
