% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runningAverage.R
\name{.gcRunningAverage}
\alias{.gcRunningAverage}
\title{Calculate GC running average}
\usage{
.gcRunningAverage(x, size = NULL)
}
\arguments{
\item{x}{A DNA sequence (a character vector). Valid bases are ACGT-.}

\item{size}{Optional. The number of observations in each average, which
can range from \code{1} to \code{length(x)}.
If not specified, the size will be set to the nearest positive, nonzero
integer to \code{length(x)/100}.}
}
\value{
A tibble with position and running average of x.
}
\description{
\code{.gcRunningAverage} calculates a running average of
the GC-content of a DNA sequence
}
\seealso{
gcContent
}
\keyword{internal}
