% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOligos.R
\name{.generateOligos}
\alias{.generateOligos}
\title{Generate oligos of a specific length}
\usage{
.generateOligos(x, oligoLength = 20, maxGapFrequency = 0.1, maxDegeneracy = 4)
}
\value{
A tibble with oligos.
}
\description{
Generate oligos of a specific length
}
\keyword{internal}
