% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{.runningAverage}
\alias{.runningAverage}
\title{Calculate running average}
\usage{
.runningAverage(x, size = NULL)
}
\arguments{
\item{x}{A numeric vector.}

\item{size}{The number of observations in each average.
If \code{NULL}, the size will be set to the nearest positive, nonzero
integer to \code{length(x) / 100}.}
}
\value{
A data frame with position and running average of \code{x}.
}
\description{
Calculate running average
}
