% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_rprimer_alignment}
\alias{example_rprimer_alignment}
\title{An alignment of hepatitis E virus sequences}
\format{
A list with class attribute 'rprimer_alignment'.
}
\source{
The sequences were collected from NCBI GenBank,
by quering “[a-z]”[porgn:__txid1678143].
Only sequences with more than 6000 bases were included.
\url{https://www.ncbi.nlm.nih.gov/}
}
\usage{
example_rprimer_alignment
}
\description{
An alignment with 621 hepatitis E virus sequences.
}
\keyword{datasets}
