% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_rprimer_oligo}
\alias{example_rprimer_oligo}
\title{Oligos generated from hepatitis E virus sequences}
\format{
A data frame (tibble) with class attribute ´rprimer_oligo'.
\describe{
  \item{begin}{position where the oligo begins}
  \item{end}{position where the oligo ends}
  \item{length}{length of the oligo}
  \item{majority}{majority sequence}
  \item{iupac}{iupac sequence (i.e. with degenerate bases)}
  \item{majority_rc}{majority sequence, reverse complement}
  \item{iupac_rc}{iupac sequence, reverse complement}
  \item{degenerates}{number of degenerate bases}
  \item{degeneracy}{number of variants}
  \item{gc_majority}{gc-content (majority sequence), proportion}
  \item{tm_majority}{melting temperature (majority sequence),
  degrees Celcius}
  \item{pm_majority}{proportion of perfectly matching sequences,
  (to the target alignment) majority sequence}
  \item{pm_iupac}{proportion of perfectly matching sequences,
  iupac sequence}
  \item{match_matrix}{a logical matrix describing which sequences
  in the target alignment the oligo matches perfectly to}
}
}
\source{
The sequences were collected from NCBI GenBank,
by querying “[a-z]”[porgn:__txid1678143].
Only sequences with more than 6000 bases were included.
\url{https://www.ncbi.nlm.nih.gov/}
}
\usage{
example_rprimer_oligo
}
\description{
Oligos generated from hepatitis E virus sequences
}
\keyword{datasets}
