% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_rprimer_sequence_profile}
\alias{example_rprimer_sequence_profile}
\title{A sequence profile from hepatitis E virus sequences}
\format{
A numeric matrix with class attribute 'rprimer_sequence_profile'.
}
\source{
The sequences were collected from NCBI GenBank,
by quering “[a-z]”[porgn:__txid1678143].
Only sequences with more than 6000 bases were included.
\url{https://www.ncbi.nlm.nih.gov/}
}
\usage{
example_rprimer_sequence_profile
}
\description{
A sequence profile from hepatitis E virus sequences
}
\keyword{datasets}
