% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_rprimer_sequence_properties}
\alias{example_rprimer_sequence_properties}
\title{Sequence properties of hepatitis E virus sequences}
\format{
A data frame (tibble) with class attribute ´
rprimer_sequence_properties'.
\describe{
  \item{position}{position in the alignment}
  \item{majority}{majority consensus sequence}
  \item{iupac}{iupac consensus sequence}
  \item{gaps}{proportion of gaps}
  \item{identity}{proportion of the most common nucleotide}
  \item{entropy}{Shannon entropy}
}
}
\source{
The sequences were collected from NCBI GenBank,
by quering “[a-z]”[porgn:__txid1678143].
Only sequences with more than 6000 bases were included.
\url{https://www.ncbi.nlm.nih.gov/}
}
\usage{
example_rprimer_sequence_properties
}
\description{
Sequence properties of hepatitis E virus sequences
}
\keyword{datasets}
