% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleRprimerAlignment}
\alias{exampleRprimerAlignment}
\title{An alignment of hepatitis E virus sequences}
\format{
A \code{Biostrings::DNAMultipleAlignment} object.
}
\source{
The sequences were collected from NCBI GenBank and the alignment
was imported to the package using \code{Biostrings::readDNAStringSet()}.
}
\usage{
data("exampleRprimerAlignment")
}
\description{
An alignment with 100 hepatitis E virus sequences.
}
\references{
H. Pages, P. Aboyoun, R. Gentleman and S. DebRoy (2020). Biostrings:
Efficient manipulation of biological strings. R package version
2.57.2.
}
\keyword{datasets}
