% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleRprimerMatchAssay}
\alias{exampleRprimerMatchAssay}
\title{Proportion of sequences that matches to assays}
\format{
An example of an \code{RprimerMatchAssay} object.
}
\usage{
data("exampleRprimerMatchAssay")
}
\description{
This dataset describes the proportion of HEV target sequences that
matches to a set of assays.
}
\keyword{datasets}
