% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleRprimerProperties}
\alias{exampleRprimerProperties}
\title{Sequence properties of hepatitis E virus sequences}
\format{
An RprimerProperties object.

\describe{
  \item{Position}{position in the alignment}
  \item{Majority}{majority consensus sequence}
  \item{IUPAC}{IUPAC consensus sequence}
  \item{Gaps}{proportion of gaps}
  \item{Identity}{proportion of the most common nucleotide}
  \item{Entropy}{Shannon entropy}
}
}
\source{
The sequences were collected from NCBI GenBank.
}
\usage{
data(exampleRprimerProperties)
}
\description{
Sequence properties of hepatitis E virus sequences
}
\keyword{datasets}
