% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignmentProperties.R
\name{gapFrequency}
\alias{gapFrequency}
\title{Gap frequency}
\usage{
gapFrequency(x)
}
\arguments{
\item{x}{An RprimerProfile object.}
}
\value{
The gap frequency (a numeric vector of length n).
}
\description{
\code{gapFrequency} returns the gap frequency from an
PrprimerProfile object.
}
\keyword{internal}
