% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_content.R
\name{gc_content}
\alias{gc_content}
\title{Calculate GC content of a DNA sequence}
\usage{
gc_content(x)
}
\arguments{
\item{x}{a DNA sequence (a character vector of length one).}
}
\value{
the GC content of x. Gaps ('-') will not be included
in the calculation.
}
\description{
\code{gc_content} finds the GC content of a DNA sequence.
}
\details{
\code{x} cannot contain other characters than
'a', 'c', 'g', 't' and '-'.
}
\examples{
gc_content("acgttcc")
gc_content("acgttcc--")
gc_content("acgrn") ## Will return an error because of an invalid base.

}
