% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcContent.R
\name{gcContent}
\alias{gcContent}
\title{Calculate GC content of a DNA sequence}
\usage{
gcContent(x)
}
\arguments{
\item{x}{A DNA sequence (a character vector of length one).
Valid bases are ACGT-.}
}
\value{
The GC content of x.
}
\description{
\code{gcContent} finds the GC content of a DNA sequence.
}
\keyword{internal}
