% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assays.R
\name{get_assays}
\alias{get_assays}
\title{Get (RT)-PCR assays from oligos}
\usage{
get_assays(x, length = 65:120, max_tm_difference = 1)
}
\arguments{
\item{x}{An object of class 'rprimer_oligo'.}

\item{length}{Amplicon length. Can range from 40 to 5000 base pairs. The
default is \code{65:120}.}

\item{max_tm_difference}{Maximum Tm difference (in C) between the two primers
(absolute value). A number between 0 and 30. The default is 1.}
}
\value{
A tibble (a data frame) of class 'rprimer_assay' with all candidate
assays. An error message will return if no assays are found.

The tibble contains the following information:

 \describe{
  \item{begin}{position where the assay begins}
  \item{end}{position where the assay ends}
  \item{amplicon_length}{length of the amplicon}
  \item{tm_difference_primer}{difference in melting temperature between
  the forward and reverse primer, absolute value, degrees Celcius}
  \item{total_degeneracy}{total number of oligos in the assay}
  \item{begin_fwd}{position where the forward primer begins}
  \item{end_fwd}{position where the reverse primer ends}
  \item{length_fwd}{length of the forward primer}
  \item{majority_fwd}{majority sequence of the forward primer}
  \item{iupac_fwd}{iupac sequence (i.e. with degenerate bases)
  of the forward primer}
  \item{degenerates_fwd}{number of degenerate bases of the forward primer}
  \item{degeneracy_fwd}{number of variants of the forward primer}
  \item{gc_majority_fwd}{gc-content of the forward primer
  (majority sequence), proportion}
  \item{tm_majority_fwd}{melting temperature of the forward primer
  (majority sequence), degrees Celcius}
  \item{pm_majority_fwd}{proportion of sequences in the target alignment
  that matches perfectly with the forward primer, majority sequence}
  \item{pm_iupac_fwd}{proportion of sequences in the target alignment
  that matches perfectly with the forward primer, iupac sequence}
  \item{begin_rev}{position where the reverse primer begins}
  \item{end_rev}{position where the reverse primer ends}
  \item{length_rev}{length of the reverse primer}
  \item{majority_rev}{majority sequence of the reverse primer}
  \item{iupac_rev}{iupac sequence (i.e. with degenerate bases)
  of the reverse primer}
  \item{degenerates_rev}{number of degenerate bases of the reverse primer}
  \item{degeneracy_rev}{number of variants of the reverse primer}
  \item{gc_majority_rev}{gc-content of the reverse primer
  (majority sequence), proportion}
  \item{tm_majority_rev}{melting temperature of the reverse primer
  (majority sequence), degrees Celcius}
  \item{pm_majority_rev}{proportion of sequences in the target alignment
  that matches perfectly with the reverse primer, majority sequence}
  \item{pm_iupac_rev}{proportion of sequences in the target alignment
  that matches perfectly with the reverse primer, iupac sequence}
  \item{pm_majority_all}{proportion of sequences in the target alignment
  that matches perfectly with both the forward and reverse primer,
  majority sequence}
  \item{pm_iupac_all}{proportion of sequences in the target alignment
  that matches perfectly with both the forward and reverse primer,
  iupac sequence}
  \item{match_matrix}{a logical matrix describing which sequences
  in the target alignment the assay matches perfectly to}
}
}
\description{
\code{get_assays} combines forward and reverse primers to (RT)-PCR assays.
}
\details{
The Tm-difference is calculated from the majority oligos, and
may thus be misleading for degenerate (iupac) oligos.
}
\examples{
get_assays(example_rprimer_oligo, length = 60:150, max_tm_difference = 1.5)

}
