% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignmentProfile.R
\name{getAlignmentProfile}
\alias{getAlignmentProfile}
\title{Get the sequence profile of an alignment}
\usage{
getAlignmentProfile(x)
}
\arguments{
\item{x}{An alignment of DNA sequences.}
}
\value{
The sequence profile (an object of class RprimerProfile).
A numeric m x n matrix,
where m is the number of unique bases in the alignment, and n is the
number of positions in the alignment.
}
\description{
\code{getAlignmentProfile} returns a matrix with the
proportion of each nucleotide at each position within an alignment
of DNA sequences.The function is a wrapper around
Biostrings::consensusMatrix.
}
\examples{
data("exampleRprimerAlignment")
getAlignmentProfile(exampleRprimerAlignment)
}
\references{
H. Pagès, P. Aboyoun, R. Gentleman and S. DebRoy (2020). Biostrings:
Efficient manipulation of biological strings. R package version 2.57.2.
}
