% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignmentProperties.R
\name{getAlignmentProperties}
\alias{getAlignmentProperties}
\title{Get alignment properties}
\usage{
getAlignmentProperties(x, iupacThreshold = 0)
}
\arguments{
\item{x}{An \code{RprimerProfile} object.}

\item{iupacThreshold}{A number [0, 0.2].
At each position, all nucleotides with a proportion
higher than the \code{iupacThreshold} will be included in
the IUPAC consensus sequence. It defaults to 0.}
}
\value{
A tibble (a data frame) with the following information:

\describe{
  \item{position}{Position in the alignment.}
  \item{majority}{Majority consensus sequence.
  The most frequently occurring nucleotide.
  If two or more bases occur with the same frequency,
  the consensus nucleotide will be randomly selected among these bases.}
  \item{iupac}{
  The consensus sequence expressed in IUPAC format (i.e. with wobble bases)
  Note that the IUPAC consensus sequence only
  takes 'A', 'C', 'G', 'T' and '-' as input. Degenerate bases
  present in the alignment will be skipped. If a position only contains
  degenerate/invalid bases, the IUPAC consensus will be \code{NA} at that
  position.}
  \item{gaps}{Proportion of gaps. Gaps are recognized as "-".}
  \item{identity}{Proportion of the most common nucleotide.
   Gaps (-), as well as bases other than A, C, G and T are excluded from the
   calculation.}
  \item{entropy}{Shannon entropy.
   Shannon entropy is a measurement of
   variability.
   First, for each nucleotide that occurs at a specific position,
   \code{p*log2(p)}, is calculated, where \code{p} is the proportion of
   that nucleotide. Then, these values are summarized,
   and multiplied by \code{-1}.
   A value of \code{0} indicate no variability and a high value
   indicate high variability.
   Gaps (-), as well as bases other than
   A, C, G and T are excluded from the calculation.}
}
}
\description{
\code{getAlignmentProperties()} returns sequence information from
an \code{RprimerProfile} object.
}
\examples{
data("exampleRprimerProfile")
getAlignmentProperties(exampleRprimerProfile)
getAlignmentProperties(exampleRprimerProfile, iupacThreshold = 0.1)

}
\seealso{
getAlignmentProfile
}
