% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConsensusProfile.R
\name{getConsensusProfile}
\alias{getConsensusProfile}
\title{Get sequence information from an alignment}
\usage{
getConsensusProfile(x, iupacThreshold = 0)
}
\arguments{
\item{x}{A \code{Biostrings::DNAMultipleAlignment} object.}

\item{iupacThreshold}{A number [0, 0.2].
At each position, all nucleotides with a proportion
higher than the \code{iupacThreshold} will be included in
the IUPAC consensus sequence. Defaults to 0.}
}
\value{
An \code{RprimerProfile} object, which contains the following information:

\describe{
  \item{position}{Position in the alignment. Note that masked columns
  from the original alignment are removed, and
  hence not taken into account when position is determined.}
  \item{a}{Proportion of A.}
  \item{c}{Proportion of C.}
  \item{g}{Proportion of G.}
  \item{t}{Proportion of T.}
  \item{other}{Proportion of bases other than A, C, G, T.}
  \item{gaps}{proportion of gaps (recognized as "-" in the alignment).}
  \item{majority}{Majority consensus sequence.
  The most frequently occurring nucleotide.
  If two or more bases occur with the same frequency,
  the consensus nucleotide will be randomly selected among these bases.}
  \item{identity}{Proportion of the most common nucleotide.
  Gaps (-), as well as bases other than A, C, G and T are excluded from the
  calculation.}
  \item{iupac}{
  The consensus sequence expressed in IUPAC format (i.e. with wobble bases)
  Note that the IUPAC consensus sequence only
  takes 'A', 'C', 'G', 'T' and '-' as input. Degenerate bases
  present in the alignment will be skipped. If a position only contains
  degenerate/invalid bases, the IUPAC consensus will be \code{NA} at that
  position.}
  \item{entropy}{Shannon entropy.
  Shannon entropy is a measurement of
  variability.
  First, for each nucleotide that occurs at a specific position,
  \code{p*log2(p)}, is calculated, where \code{p} is the proportion of
  that nucleotide. Then, these values are summarized,
  and multiplied by \code{-1}.
  A value of \code{0} indicate no variability and a high value
  indicate high variability.
  Gaps (-), as well as bases other than
  A, C, G and T are excluded from the calculation.}
}
}
\description{
\code{getConsensusProfile()} returns sequence information from an alignment
of DNA sequences.
}
\examples{
data("exampleRprimerAlignment")
getConsensusProfile(exampleRprimerAlignment)
}
\references{
This function is a wrapper around \code{Biostrings::consensusMatrix()}:

H. Pagès, P. Aboyoun, R. Gentleman and S. DebRoy (2020). Biostrings:
Efficient manipulation of biological strings. R package version
2.57.2.
}
