% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_rprimer_alignment.R
\name{is.rprimer_alignment}
\alias{is.rprimer_alignment}
\title{Check if an object is an rprimer_alignment}
\usage{
is.rprimer_alignment(x)
}
\description{
Check if an object is an rprimer_alignment
}
