% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{is.RprimerProfile}
\alias{is.RprimerProfile}
\title{Check if an object is as RprimerProfile-object}
\usage{
is.RprimerProfile(x)
}
\arguments{
\item{x}{An 'RprimerProfile'-like object.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Check if an object is as RprimerProfile-object
}
