% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignmentProperties.R
\name{iupacConsensus}
\alias{iupacConsensus}
\title{IUPAC consensus sequence}
\usage{
iupacConsensus(x, threshold = 0)
}
\arguments{
\item{x}{An RprimerProfile object.}

\item{threshold}{Optional. A number (0, 0.2]
At each position, all nucleotides with a proportion
higher or equal to the threshold will be included in
the IUPAC consensus sequence. The default is 0.}
}
\value{
The consensus sequence (a character vector of length n).
}
\description{
\code{iupacConsensus} returns the IUPAC consensus sequence from an
PrprimerProfile object.
}
\seealso{
\code{asIUPAC} for further info on how the IUPAC consensus
sequecnce is determined.
}
\keyword{internal}
