% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignmentProperties.R
\name{majorityConsensus}
\alias{majorityConsensus}
\title{Majority consensus sequence}
\usage{
majorityConsensus(x)
}
\arguments{
\item{x}{An RprimerProfile object.}
}
\value{
The majority consensus sequence (a character vector of length n).
}
\description{
\code{majorityConsensus} returns the majority consensus sequence of an
alignment of DNA sequences.
}
\keyword{internal}
