% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignmentProperties.R
\name{nucleotideIdentity}
\alias{nucleotideIdentity}
\title{Nucleotide identity}
\usage{
nucleotideIdentity(x)
}
\arguments{
\item{x}{An RprimerProfile object.}
}
\value{
The nucleotide identity (a numeric vector of length n).
The nucleotide identity can range between (0, 1].
}
\description{
\code{nucleotideIdentity} returns the nucleotide identity from an
PrprimerProfile object.
}
\keyword{internal}
