% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oligos.R
\name{oligos}
\alias{oligos}
\title{Design oligos}
\usage{
oligos(
  x,
  maxGapFrequency = 0.1,
  lengthPrimer = 18:22,
  maxDegeneracyPrimer = 4,
  gcClampPrimer = TRUE,
  avoidThreeEndRunsPrimer = TRUE,
  minEndIdentityPrimer = 0,
  gcRangePrimer = c(0.45, 0.55),
  tmRangePrimer = c(55, 65),
  concPrimer = 500,
  probe = TRUE,
  lengthProbe = 18:22,
  maxDegeneracyProbe = 4,
  avoidFiveEndGProbe = TRUE,
  gcRangeProbe = c(0.45, 0.55),
  tmRangeProbe = c(55, 70),
  concProbe = 250,
  concNa = 0.05
)
}
\arguments{
\item{x}{An \code{RprimerProfile} object.}

\item{maxGapFrequency}{Maximum allowed gap frequency (for both primers and probes).
A number [0, 1], defaults to 0.1.}

\item{lengthPrimer}{Primer length. A numeric vector [14, 30],
defaults to \code{18:22}.}

\item{maxDegeneracyPrimer}{Maximum number of variants of each primer. A number [1, 32], defaults to 4.}

\item{gcClampPrimer}{If primers with no GC-clamp
should be avoided.
A GC-clamp
is identified as two to three G or
C:s within the last five bases (3'-end) of the oligo.
\code{TRUE} or \code{FALSE}, defaults to \code{TRUE}.}

\item{avoidThreeEndRunsPrimer}{If primers with more than two runs
of the same nucleotide at the terminal 3'-end should be avoided.
\code{TRUE} or \code{FALSE}, defaults to \code{TRUE}.}

\item{minEndIdentityPrimer}{The minimum allowed identity
at the 3'-end of the primer (the last five bases).
A number [0, 1], defaults to 0.}

\item{gcRangePrimer}{GC-content range for primers (proportion, not percent).
A numeric vector [0, 1], defaults to \code{c(0.45, 0.55)}.}

\item{tmRangePrimer}{Tm range for primers.
A numeric vector [30, 90], defaults to \code{c(55, 65)}.}

\item{concPrimer}{Primer concentration in nM, for Tm calculation. A number
[20, 2000], defaults to 500.}

\item{probe}{If probes should be designed. \code{TRUE} or \code{FALSE},
defaults to \code{TRUE}.}

\item{lengthProbe}{Probe length. A numeric vector [14, 30],
defaults to \code{18:22}.}

\item{maxDegeneracyProbe}{Maximum number of variants of each probe. A number [1, 32], defaults to 4.}

\item{avoidFiveEndGProbe}{If probes with G
at the 5'-end should be avoided. \code{TRUE} or \code{FALSE},
defaults to \code{TRUE}.}

\item{gcRangeProbe}{GC-content range for probes (proportion, not %). A numeric vector [0, 1],
defaults to \code{c(0.45, 0.55)}.}

\item{tmRangeProbe}{Tm range for probes.
A numeric vector [30, 90], defaults to \code{c(55, 65)}.}

\item{concProbe}{Primer concentration in nM, for Tm calculation. A numeric vector
[20, 2000], defaults to 250.}

\item{concNa}{The sodium ion concentration in the PCR reaction in M, for Tm calculation.
A numeric vector [0.01, 1], defaults to 0.05 (50 mM).}
}
\value{
An \code{RprimerOligo} object. An error message will return
if no oligos are found.
}
\description{
\code{oligos()} designs oligos (primers and probes)
from an \code{RprimerProfile} object.
}
\section{Output}{


The output contains the following information:

\describe{
  \item{type}{Whether the oligo is a primer or probe.}
  \item{fwd}{\code{TRUE} if the oligo is valid in forward direction,
    \code{FALSE} otherwise.}
  \item{rev}{\code{TRUE} if the oligo is valid in reverse direction,
    \code{FALSE} otherwise.}
  \item{start}{Start position of the oligo.}
  \item{end}{End positon of the oligo.}
  \item{length}{Oligo length.}
  \item{iupacSequence}{Oligo sequence, with wobble bases (if any).}
  \item{iupaSequenceRc}{The reverse complement of the iupacSequence.}
  \item{identity}{Average identity score of the oligo, can range from 0 to 1.
    The identity is the proportion of the most common base at each position
    in the input alignment.}
  \item{degeneracy}{Number of sequence variants of the oligo.}
  \item{gcContentMean}{Mean GC-content of all sequence variants of the oligo.
  }
  \item{gcContentRange}{Range in GC-content of all sequence variants of
    the oligo.}
  \item{tmMean}{Mean tm of all sequence variants of the oligo.}
  \item{tmRange}{Range in tm of all sequence variants of the oligo.}
  \item{sequence}{All sequence variants of the oligo.}
  \item{sequenceRc}{Reverse complements of all sequence variants.}
  \item{gcContent}{GC-content of all sequence variants.}
  \item{tm}{tm of all sequence variants.}
  \item{roiStart}{First position of the input \code{RprimerProfile} object
    (roi = region of interest).}
  \item{roiEnd}{Last position of the input \code{RprimerProfile} object.}
}
}

\section{Description of the design process}{


to be added.
}

\section{Tm-calculation}{


Melting temperatures are calculated using SantaLucia's nearest-neighbor
method, with the following assumptions:

\itemize{
  \item Oligos are not expected to be self-complementary (no symmetry
  correction is done).
  \item The oligo concentration is assumed to be much higher
  than the target concentration.
}

See references for table values and equations. Table values can also be
found by calling \code{rprimer:::lookup$nn}.
}

\examples{
data("exampleRprimerProfile")
target <- exampleRprimerProfile

## Design primers and probes with default values
oligos(target)

## Design primers and probes only within a specific region of interest
roi <- target[target$position >= 5000 & target$position <= 6000, ]
oligos(roi)

## Design primers only
oligos(roi, probe = FALSE)

## Allow high degeneracy, but enforce conserved 3'-ends (>= 98 \% identity)
oligos(roi,
       maxDegeneracyPrimer = 8,
       minEndIdentityPrimer = 0.98,
       probe = FALSE
)
}
\references{
SantaLucia Jr, J., & Hicks, D. (2004).
The thermodynamics of DNA structural motifs.
Annu. Rev. Biophys. Biomol. Struct., 33, 415-440.
}
