% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotAlignmentOverview}
\alias{plotAlignmentOverview}
\alias{plotAlignmentOverview,RprimerProfile-method}
\title{Plot an overview of an RprimerProfile object.}
\usage{
plotAlignmentOverview(x)

\S4method{plotAlignmentOverview}{RprimerProfile}(x)
}
\arguments{
\item{x}{An \code{RprimerProfile} object.}
}
\value{
A plot.
}
\description{
Plot an overview of an RprimerProfile object.
}
\section{Methods (by class)}{
\itemize{
\item \code{RprimerProfile}: 
}}

