% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotConsensusProfile}
\alias{plotConsensusProfile}
\title{Plot sequence conservation, GC content and gap frequency}
\usage{
plotConsensusProfile(x, shadeFrom = NULL, shadeTo = NULL)
}
\arguments{
\item{x}{An \code{RprimerProfile} object.}

\item{shadeFrom}{Optional. If a particular area should be shaded, where the shade
should start (an integer).}

\item{shadeTo}{Optional. If a particular area should be shaded, where the shade
should end (an integer).}
}
\value{
A plot.
}
\description{
Plot sequence conservation, GC content and gap frequency
}
