% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotNucleotides}
\alias{plotNucleotides}
\alias{plotNucleotides,RprimerProfile-method}
\title{Plot the nucleotide distribution of an RprimerProfile object}
\usage{
plotNucleotides(x, rc = FALSE)

\S4method{plotNucleotides}{RprimerProfile}(x, rc = FALSE)
}
\arguments{
\item{x}{An \code{RprimerProfile} object.}

\item{rc}{\code{TRUE} or {FALSE}. If the plotted sequence should be displayed
as reverse complement or not. Defaults to \code{FALSE}.}
}
\value{
A barplot with nucleotide distribution.
}
\description{
Plot the nucleotide distribution of an RprimerProfile object
}
\section{Methods (by class)}{
\itemize{
\item \code{RprimerProfile}: 
}}

\examples{
data("exampleRprimerProfile")
plotNucleotides(exampleRprimerProfile[, 1:10]) ## Plot the first 10 bases
plotNucleotides(exampleRprimerProfile[, 1:10], rc = TRUE)
## As reverse complement

}
