% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_complement.R
\name{reverse_complement}
\alias{reverse_complement}
\title{Reverse complement}
\usage{
reverse_complement(x)
}
\arguments{
\item{x}{A DNA sequence (a character vector of length one).}
}
\value{
The reverse complement. Non valid bases will return as \code{NA}.
}
\description{
\code{reverse_complement} finds the reverse complement of a DNA seuquence.
}
\details{
For \code{x}, valid bases are 'a', 'c', 'g', 't', 'r', 'y', 'm',
'k', 's', 'w', n', 'h', 'd', 'v', 'b' and '-'.
}
\examples{
reverse_complement("cttgtr")

}
