% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseComplement.R
\name{reverseComplement}
\alias{reverseComplement}
\title{Reverse complement}
\usage{
reverseComplement(x)
}
\arguments{
\item{x}{A DNA sequence (a character vector of length one). Valid bases are
ACGTRYSWKMBDHVN-.}
}
\value{
The reverse complement of x. Non valid bases will return as \code{NA}.
}
\description{
\code{reverseComplement} finds the reverse complement of a DNA sequence.
}
\keyword{internal}
