% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_exported_functions.R
\name{round_df_dbl}
\alias{round_df_dbl}
\title{Round all doubles in a data frame}
\usage{
round_df_dbl(x)
}
\description{
@param x A data frame/tibble.
}
\details{
@return A data frame/tibble where all vectors
 of type 'double' have been rounded to two digits.

 @noRd
}
