% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprimerProfile.R
\name{rpGetData,RprimerProfile-method}
\alias{rpGetData,RprimerProfile-method}
\title{Getter, method for RprimerProfile objects}
\usage{
\S4method{rpGetData}{RprimerProfile}(x)
}
\arguments{
\item{x}{An \code{RprimerProfile} object.}
}
\description{
Getter, method for RprimerProfile objects
}
