% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprimerProfile.R
\name{rpGetData}
\alias{rpGetData}
\title{Getter, generic}
\usage{
rpGetData(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\value{
A matrix.
}
\description{
Getter, generic
}
