% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpPlot.R
\name{rpPlot}
\alias{rpPlot}
\alias{rpPlot,RprimerProfile-method}
\title{Plot an rprimer object (generic)}
\usage{
rpPlot(x, ...)

\S4method{rpPlot}{RprimerProfile}(x, rc = FALSE)
}
\arguments{
\item{x}{An rprimer object (see methods for details).}

\item{...}{Additional arguments that should be passed to the plot, when
plotting an RprimerProfile object.}

\item{rc}{\code{TRUE/FALSE}. If the plotted sequence should be displayed
as reverse complement or not. The default is \code{FALSE}.}
}
\value{
A plot.
}
\description{
Plot an rprimer object (generic)
}
\section{Methods (by class)}{
\itemize{
\item \code{RprimerProfile}: Plot an RprimerProfile object.
}}

\examples{

}
