% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RprimerAssay-class}
\alias{RprimerAssay-class}
\alias{.RprimerAssay}
\alias{RprimerAssay}
\title{An S4 class for representation of PCR assays}
\usage{
RprimerAssay(...)
}
\arguments{
\item{...}{A data frame or list to be converted into an \code{RprimerAssay}
object.}
}
\value{
An \code{RprimerAssay} object if validation succeeds, an error
message otherwise.
}
\description{
\code{RprimerAssay} extends the \code{S4Vectors::DataFrame} class,
without any additional slots. It has the same accessors and
coercion, subsetting, and combining methods as the parent class, but has
some additional checks for validity.

The constructor, \code{RprimerAssay()},
constructs an \code{RprimerAssay} object in a similar fashion as the
\code{S4Vectors::DataFrame()} constructor.
}
\section{Functions}{
\itemize{
\item \code{RprimerAssay}: 
}}

\examples{
data("exampleRprimerAssay")
x <- as.data.frame(exampleRprimerAssay)
RprimerAssay(x)
}
\references{
Pages, H., Lawrence, M., and Aboyoun, R. (2020). S4Vectors:
Foundation of vector-like and list-like containers in
Bioconductor. R package version 0.28.0.
}
\seealso{
Other RprimerAssay: 
\code{\link{coerce}()}
}
\concept{RprimerAssay}
