% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RprimerMatchOligo-class}
\alias{RprimerMatchOligo-class}
\alias{.RprimerMatchOligo}
\alias{RprimerMatchOligo}
\title{An S4 class for representation of match proportions of oligos}
\usage{
RprimerMatchOligo(...)
}
\arguments{
\item{...}{A data frame or list to be converted into an \code{RprimerMatchOligo}
object.}
}
\value{
An \code{RprimerMatchOligo} object if validation succeeds, an error
message otherwise.
}
\description{
\code{RprimerMatchOligo} extends the \code{S4Vectors::DataFrame} class,
without any additional slots. It has the same accessors and
coercion, subsetting, and combining methods as the parent class, but has
some additional checks for validity.

The constructor, \code{RprimerMatchOligo()},
constructs an \code{RprimerMatchOligo} object in a similar fashion as the
\code{S4Vectors::DataFrame()} constructor.
}
\section{Functions}{
\itemize{
\item \code{RprimerMatchOligo}: 
}}

\examples{
data("exampleRprimerMatchOligo")
x <- as.data.frame(exampleRprimerMatchOligo)
RprimerMatchOligo(x)
}
\references{
Pages, H., Lawrence, M., and Aboyoun, R. (2020). S4Vectors:
Foundation of vector-like and list-like containers in
Bioconductor. R package version 0.28.0.
}
