% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RprimerOligo-class}
\alias{RprimerOligo-class}
\alias{.RprimerOligo}
\alias{RprimerOligo}
\title{An S4 class for representation of oligos}
\usage{
RprimerOligo(..., row.names = NULL, check.names = TRUE, stringsAsFactors)
}
\arguments{
\item{...}{A data frame or list to be converted into an \code{RprimerProfile}
object.}

\item{row.names}{\code{NULL} or a single integer or character string specifying a column to be
used as row names, or a character or integer vector giving the row names
for the data frame.}

\item{check.names}{logical. If \code{TRUE} then the names of the variables in the data frame are
checked to ensure that they are syntactically valid variable names and
are not duplicated.}

\item{stringsAsFactors}{logical. If character vectors should be converted to factors.}
}
\value{
An \code{RprimerOligo} object if validation succeeds, an error
message otherwise.
}
\description{
\code{RprimerOligo} extends the \code{S4Vectors::DataFrame} class,
without any additional slots. It has the same accessors and
coercion, subsetting, and combining methods as the parent class, but has
some additional checks for validity.

Constructs an \code{RprimerOligo} object in a similar fashion as the
\code{S4Vectors::DataFrame} constructor, without any additional arguments.
}
\section{Functions}{
\itemize{
\item \code{RprimerOligo}: 
}}

\examples{
data("exampleRprimerOligo")
x <- as.data.frame(exampleRprimerOligo)
RprimerOligo(x)
}
