% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{RprimerProfile}
\alias{RprimerProfile}
\title{RprimerProfile}
\usage{
RprimerProfile(..., row.names = NULL, check.names = TRUE, stringsAsFactors)
}
\arguments{
\item{...}{A data frame or list to be converted into an \code{RprimerProfile}
object.}

\item{row.names}{\code{NULL} or a single integer or character string specifying a column to be
used as row names, or a character or integer vector giving the row names
for the data frame.}

\item{check.names}{logical. If \code{TRUE} then the names of the variables in the data frame are
checked to ensure that they are syntactically valid variable names and
are not duplicated.}

\item{stringsAsFactors}{logical. If character vectors should be converted to factors.}
}
\value{
An \code{RprimerAssay} object if validation succeeds, an error
message otherwise.
}
\description{
Constructs an \code{RprimerProfile} object in a similar fashion as the
\code{S4Vectors::DataFrame} constructor, without any additional arguments.
}
\examples{
data("exampleRprimerProfile")
x <- as.data.frame(exampleRprimerProfile)
RprimerProfile(x)

}
