% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runningAverage.R
\name{runningAverage}
\alias{runningAverage}
\title{Calculate running average}
\usage{
runningAverage(x, size = NULL)
}
\arguments{
\item{x}{A numeric vector.}

\item{size}{Optional. The number of observations in each average, which
can range from \code{1} to \code{length(x)}.
If not specified, the size will be set to the nearest positive, nonzero
integer to \code{length(x)/100}.}
}
\value{
A tibble with position and running average of x.
}
\description{
\code{runningAverage} calculates a running average of a
numeric vector.
}
\keyword{internal}
