% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRprimerApp.R
\name{runRprimerApp}
\alias{runRprimerApp}
\title{rprimer Shiny application}
\usage{
runRprimerApp()
}
\value{
Opens the Shiny application.
}
\description{
\code{runRprimerApp()} starts a Shiny application where
the workflow of the rprimer package can be run through a
graphical user interface.
}
\examples{
## Only run this in interactive R sessions:
if (interactive()) {
    runRprimerApp()
}
}
