% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{select_roi}
\alias{select_roi}
\title{Select a region of interest of an alignment}
\usage{
select_roi(x, from = 1, to = NULL)
}
\arguments{
\item{x}{An alignment of DNA sequences
(an object of class 'rprimer_alignment').}

\item{from}{Where the roi begins (an integer).}

\item{to}{Where the roi ends (an integer, greater than from).}
}
\value{
The roi of the alignment
(an object of class 'rprimer_alignment').
}
\description{
\code{select_roi} selects a specified region of interest of
an alignment of DNA sequences.
}
\details{
The roi must be at least 200 bases. Note that the positions
will
not be kept from the input alignment. The new alignment will always
start at position 1.
}
\examples{
# Select the first 1000 bases
select_roi(example_rprimer_alignment, from = 1, to = 1000)

}
\seealso{
\code{read_fasta_alignment}.
}
