% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_profile.R
\name{sequence_profile}
\alias{sequence_profile}
\title{Get the sequence profile of an alignment}
\usage{
sequence_profile(x)
}
\arguments{
\item{x}{An alignment of DNA sequences (an object of class 'rprimer_alignment').}
}
\value{
The sequence profile (an object of class rprimer_profile').
A numeric m x n matrix,
where m is the number of unique bases in the alignment, and n is the
number of positions in the alignment.
}
\description{
\code{sequence_profile} returns a matrix with the
proportion of each nucleotide at each position within an alignment
of DNA sequences.
}
\examples{
sequence_profile(example_rprimer_alignment)

}
