% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{sequence_properties}
\alias{sequence_properties}
\title{Get sequence properties}
\usage{
sequence_properties(x, iupac_threshold = 0)
}
\arguments{
\item{x}{A nucleotide profile of class 'rprimer_sequence_profile'.}

\item{iupac_threshold}{A number between 0 and 0.2 (the default is 0).
At each position, all nucleotides with a proportion
higher than or equal to the stated threshold will be included in
the iupac consensus sequence.}
}
\value{
A tibble (data frame) of class 'rprimer_sequence_properties',
with information about majority and iupac consensus sequence, gap frequency,
nucleotide identity and shannon entropy.
}
\description{
\code{sequence_properties} returns sequence information from an alignment
of DNA sequences.
}
\section{Majority consensus sequence}{

The most frequently occuring nucleotide.
If two or more bases occur with the same frequency,
the consensus nucleotide will be randomly selected among these bases.
}

\section{IUPAC consensus sequence}{

The consensus sequence expressed in IUPAC format (i.e. with wobble bases)
Note that the IUPAC consensus sequence only
takes 'a', 'c', 'g', 't' and '-' as input. Degenerate bases
present in the alignment will be skipped. If a position only contains
degenerate/invalid bases, the IUPAC consensus will be \code{NA} at that
position.
}

\section{Gaps}{

Gaps are recognised as "-" in the sequence profile.
}

\section{Identity}{

The nucleotide identity is the proportion of
the most common base. Gaps (-),
as well as nucleotides other than a, c, g and t, are excluded from the
calculation.
}

\section{Shannon entropy}{

Shannon entropy is a measurement of
variability. First, for each nucleotide that occurs at a specific position,
\code{p*log2(p)}, is calculated, where \code{p} is the proportion of
that nucleotide. Then, the shannon entropy is calculated by summarising
these values for each nucleotide at the position in matter,
followed by multiplication by \code{-1}.
A value of \code{0} indicate no variability and a high value
indicate high variability.
Gaps (-), as well as bases other than
a, c, g and t, are excluded from the calculation.
}

\examples{
sequence_properties(example_rprimer_sequence_profile)

}
