% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignmentProperties.R
\name{shannonEntropy}
\alias{shannonEntropy}
\title{Shannon entropy}
\usage{
shannonEntropy(x)
}
\arguments{
\item{x}{An RprimerProfile object.}
}
\value{
The Shannon entropy (a numeric vector of length n).
}
\description{
\code{shannonEntropy} returns the Shannon entropy from an
PrprimerProfile object.
}
\keyword{internal}
