% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_rprimer_alignment.R
\name{[.rprimer_alignment}
\alias{[.rprimer_alignment}
\title{Extract elements in an rprimer_alignment object}
\usage{
\method{[}{rprimer_alignment}(x, i, ...)
}
\arguments{
\item{x}{Object from which to extract element(s).}

\item{i}{Indices specifying elements to extract.}

\item{...}{Indices specifying elements to extract.}
}
\description{
Extract elements in an rprimer_alignment object
}
