% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_rprimer_alignment.R
\name{[<-.rprimer_alignment}
\alias{[<-.rprimer_alignment}
\title{Replace elements in an rprimer_alignment object}
\usage{
\method{[}{rprimer_alignment}(x, i) <- value
}
\arguments{
\item{x}{Object from which to replace element(s) from.}

\item{i}{Indices specifying elements to replace.}

\item{value}{Typically an array-like object of a similar class as x.}
}
\description{
Replace elements in an rprimer_alignment object
}
