% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{write_assay_report}
\alias{write_assay_report}
\title{Write assay report}
\usage{
write_assay_report(
  filename = "my_assay_report",
  assay_selection,
  alignment,
  sequence_profile,
  sequence_properties
)
}
\description{
Write assay report
}
