% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assays.R
\name{assays}
\alias{assays}
\title{Design (RT)-PCR assays}
\usage{
assays(
  x,
  lengthRange = c(65, 120),
  tmDiffPrimers = 5,
  tmDiffPrimersProbe = c(-10, 10)
)
}
\arguments{
\item{x}{An \code{RprimerOligo} object, with or without probes.}

\item{lengthRange}{Amplicon length range, a numeric vector [40, 5000], defaults to
\code{c(65, 120)}.}

\item{tmDiffPrimers}{Maximum allowed difference between the mean tm of the forward primer
and the mean tm of the reverse primer
(absolute value). A number [0, 20], defaults to 5.}

\item{tmDiffPrimersProbe}{Acceptable difference between the mean tm of the probe and the mean tm
of the primers. A numeric vector [-20, 20],
defaults to \code{c(-10, 10)}.
A negative tm difference
means that the mean tm of the probe is lower than the mean tm of the
primers.}
}
\value{
An \code{RprimerAssay} object.
An error message will return if no assays are found.

The object contains the following information:

\describe{
  \item{start}{Position where the assay starts.}
  \item{end}{Position where the assay ends.}
  \item{ampliconLength}{Length of the amplicon.}
  \item{tmDifferencePrimer}{Difference between
  the mean tm of the forward primer and the mean tm of the reverse primer,
  absolute value.}
  \item{deltaGDifferencePrimer}{Difference between
  the mean delta G of the forward primer and the mean delta G of
  the reverse primer, absolute value.}
  \item{totalDegeneracy}{Total number of oligos in the assay.}
  \item{score}{Summarized oligo score. The lowest, and best,
  possible score is 0. The highest possible score is 24 for assays
  with only primers,
  and 36 for assays with probes.
  See \code{?oligos} for more information about the scoring system.}
  \item{startFwd}{Position where the forward primer starts.}
  \item{endFwd}{Position where the forward primer ends.}
  \item{lengthFwd}{Length of the forward primer.}
  \item{iupacSequenceFwd}{IUPAC sequence of the forward primer.}
  \item{coverageFwd}{Average coverage of the forward primer.}
  \item{degeneracyFwd}{Number of variants of the forward primer.}
  \item{gcContentMeanFwd}{Mean GC-content of the forward primer.}
  \item{gcContentRangeFwd}{Range in GC-content of the forward primer.}
  \item{tmMeanFwd}{Mean tm of the forward primer.}
  \item{tmRangeFwd}{Range in tm of the forward primer.}
  \item{deltaGMeanFwd}{Mean delta G of the forward primer.}
  \item{deltaGRangeFwd}{Range in delta G of the forward primer.}
  \item{sequenceFwd}{Sequence of the forward primer, all variants.}
  \item{gcContentFwd}{GC-content of the forward primer, all variants.}
  \item{tmFwd}{Tm of the forward primer, all variants.}
  \item{deltaGFwd}{Delta G of the forward primer, all variants.}
  \item{startRev}{Position where the reverse primer starts.}
  \item{endRev}{Position where the reverse primer ends.}
  \item{lengthRev}{Length of the reverse primer.}
  \item{iupacSequenceRev}{IUPAC sequence of the reverse primer.}
  \item{coverageRev}{Average coverage of the reverse primer.}
  \item{degeneracyRev}{Number of variants of the reverse primer.}
  \item{gcContentMeanRev}{Mean GC-content of the reverse primer.}
  \item{gcContentRangeRev}{Range in GC-content of the reverse primer.}
  \item{tmMeanRev}{Mean tm of the reverse primer.}
  \item{tmRangeRev}{Range in tm of the reverse primer.}
  \item{deltaGMeanRev}{Mean delta G of the reverse primer.}
  \item{deltaGRangeRev}{Range in delta G of the reverse primer.}
  \item{sequenceRev}{Sequence of the reverse primer, all variants.}
  \item{gcContentRev}{GC-content of the reverse primer, all variants.}
  \item{tmRev}{Tm of the reverse primer, all variants.}
  \item{deltaGRev}{Delta G of the reverse primer, all variants.}
  \item{roiStart}{Start position of the input consensus profile
  used for oligo design.}
  \item{roiEnd}{End position of the input consensus profile used
  for oligo design.}
}

If a probe is used, the following columns are also included:

\describe{
  \item{tmDifferencePrimerProbe}{Tm difference between the primer pair
  and probe. The tm difference is calculated by subtracting the
  mean tm of the probe with the mean tm of the primers.}
  \item{deltaGDifferencePrimerProbe}{Difference in delta G between
  the primer pair and probe. It is calculated by subtracting the
  mean delta G of the probe with the mean delta G of the primers.}
  \item{plusPr}{If the probe is valid in positive sense.}
  \item{minusPr}{If the probe is valid in negative sense.}
  \item{startPr}{Position where the probe starts.}
  \item{endPr}{Position where the probe ends.}
  \item{lengthPr}{Length of the probe.}
  \item{iupacSequencePr}{IUPAC sequence of the probe.}
  \item{coveragePr}{Average coverage of the probe.}
  \item{degeneracyPr}{Number of variants of the probe.}
  \item{gcContentMeanPr}{Mean GC-content of the probe.}
  \item{gcContentRangePr}{Range in GC-content of the probe.}
  \item{tmMeanPr}{Mean tm of the probe.}
  \item{tmRangePr}{Range in tm of the probe.}
  \item{deltaGMeanPr}{Mean delta G of the probe.}
  \item{deltaGRangePr}{Range in delta G of the probe.}
  \item{sequencePr}{Sequence of the probe, all variants.}
  \item{gcContentPr}{GC-content of the probe, all variants.}
  \item{tmPr}{Tm of the probe, all variants.}
  \item{deltaGPr}{Delta G of the probe, all variants.}
}
}
\description{
\code{assays()} combines forward, reverse primers
and probes to (RT)-PCR assays from an \code{RprimerOligo} object.
}
\examples{
data("exampleRprimerOligo")
## Get assays using default settings
assays(exampleRprimerOligo)
}
