% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assays.R
\name{assays}
\alias{assays}
\title{Design (RT)-PCR assays}
\usage{
assays(
  x,
  lengthRange = c(65, 120),
  tmDiffPrimers = 2,
  tmDiffPrimersProbe = c(0, 20)
)
}
\arguments{
\item{x}{An \code{RprimerOligo} object, with or without probes.}

\item{lengthRange}{Amplicon length range, a numeric vector [40, 5000], defaults to
\code{c(65, 120)}.}

\item{tmDiffPrimers}{Maximum tm difference between the two primers
(absolute value). A number [0, 30], defaults to 2.}

\item{tmDiffPrimersProbe}{Acceptable tm difference between the primers (average Tm of the
primer pair) and probe. A numeric vector [-20, 20],
defaults to \code{c(0, 20)}.
The Tm-difference is calculated by subtracting the
Tm of the probe with the average Tm of the majority
primer pair. A negative Tm-difference
means that the Tm of the probe is lower than the average Tm of the
primer pair.}
}
\value{
An \code{RprimerAssay} object.
An error message will return if no assays are found.

The object contains the following information:

\describe{
  \item{start}{Position where the assay starts.}
  \item{end}{Position where the assay ends.}
  \item{ampliconLength}{Length of the amplicon.}
  \item{tmDifferencePrimer}{Difference in Tm between
  the forward and reverse primer, absolute value.}
  \item{totalDegeneracy}{Total number of oligos in the assay.}
  \item{startFwd}{Position where the forward primer starts.}
  \item{endFwd}{Position where the forward primer ends.}
  \item{lengthFwd}{Length of the forward primer.}
  \item{iupacSequenceFwd}{IUPAC sequence of the forward primer.}
  \item{identityFwd}{Average identity of the forward primer.}
  \item{degeneracyFwd}{Number of variants of the forward primer.}
  \item{gcContentMeanFwd}{Mean GC-content of the forward primer.}
  \item{gcContentRangeFwd}{Range in GC-content of the forward primer.}
  \item{tmMeanFwd}{Mean tm of the forward primer.}
  \item{tmRangeFwd}{Range in tm of the forward primer.}
  \item{sequenceFwd}{Sequence of the forward primer, all variants.}
  \item{gcContentFwd}{GC-content of the forward primer, all variants.}
  \item{tmFwd}{Tm of the forward primer, all variants.}
  \item{startRev}{Position where the reverse primer starts.}
  \item{endRev}{Position where the reverse primer ends.}
  \item{lengthRev}{Length of the reverse primer.}
  \item{iupacSequenceRev}{IUPAC sequence of the reverse primer.}
  \item{identityRev}{Average identity of the reverse primer.}
  \item{degeneracyRev}{Number of variants of the reverse primer.}
  \item{gcContentMeanRev}{Mean GC-content of the reverse primer.}
  \item{gcContentRangeRev}{Range in GC-content of the reverse primer.}
  \item{tmMeanRev}{Mean tm of the reverse primer.}
  \item{tmRangeRev}{Range in tm of the reverse primer.}
  \item{sequenceRev}{Sequence of the reverse primer, all variants.}
  \item{gcContentRev}{GC-content of the reverse primer, all variants.}
  \item{tmRev}{Tm of the reverse primer, all variants.}
  \item{roiStart}{Start position of the input consensus profile
  used for oligo design.}
  \item{roiEnd}{End position of the input consensus profile used
  for oligo design.}
}

If a probe is used, the following columns are also included:

\describe{
  \item{plusPr}{If the probe is valid in positive sense.}
  \item{minusPr}{If the probe is valid in negative sense.}
  \item{startPr}{Position where the probe starts.}
  \item{endPr}{Position where the probe ends.}
  \item{lengthPr}{Length of the probe.}
  \item{iupacSequencePr}{IUPAC sequence of the probe.}
  \item{identityPr}{Average identity of the probe.}
  \item{degeneracyPr}{Number of variants of the probe.}
  \item{gcContentMeanPr}{Mean GC-content of the probe.}
  \item{gcContentRangePr}{Range in GC-content of the probe.}
  \item{tmMeanPr}{Mean tm of the probe.}
  \item{tmRangePr}{Range in tm of the probe.}
  \item{sequencePr}{Sequence of the probe, all variants.}
  \item{gcContentPr}{GC-content of the probe, all variants.}
  \item{tmPr}{Tm of the probe, all variants.}
}
}
\description{
\code{assays()} combines forward and reverse primers
and probes (if selected) to (RT)-PCR assays.
}
\section{Tm-differences}{

Note that the tm-differences are calculated from the mean tm of each oligo.
}

\examples{
data("exampleRprimerOligo")
## Get assays using default settings
assays(exampleRprimerOligo)
}
