% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMatch.R
\name{checkMatch}
\alias{checkMatch}
\alias{checkMatch,RprimerOligo-method}
\title{Check how oligos match to a set of target sequences (generic)}
\usage{
checkMatch(x, target)

\S4method{checkMatch}{RprimerOligo}(x, target)
}
\arguments{
\item{x}{An \code{RprimerOligo} or \code{RprimerAssay} object.}

\item{target}{The \code{Biostrings::DNAMultipleAlignment} object used for generating the
consensus profile, with intended target sequences.

## Write details and limitations}
}
\value{
A table with oligo sequences and proportion of sequences ######################
(An \code{RprimerMatchOligo} or \code{RprimerMatchAssay} object).
}
\description{
Check how oligos match to a set of target sequences (generic)

Check match of an RprimerOligo object (method)
}
\section{Methods (by class)}{
\itemize{
\item \code{RprimerOligo}: 
}}

