% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToDNAStringSet.R
\name{convertToDNAStringSet}
\alias{convertToDNAStringSet}
\alias{convertToDNAStringSet,RprimerOligo-method}
\alias{convertToDNAStringSet,RprimerAssay-method}
\title{Convert to DNA string set (generic)}
\usage{
convertToDNAStringSet(x, asRc = FALSE, asIUPAC = TRUE)

\S4method{convertToDNAStringSet}{RprimerOligo}(x, asRc = FALSE, asIUPAC = TRUE)

\S4method{convertToDNAStringSet}{RprimerAssay}(x, asRc = FALSE, asIUPAC = TRUE)
}
\arguments{
\item{x}{An \code{RprimerOligo} or \code{RprimerAssay} object.}

\item{asRc}{For \code{RprimerOligo} objects: if oligos should be
written as reverse complements.
For \code{RprimerAssay} objects: if reverse primers and minus
sense probes should be
written in reverse complement.
Defaults to \code{FALSE}.}

\item{asIUPAC}{If sequences should be presented in IUPAC (degenerate) format. Defaults to
\code{TRUE}. If set to \code{FALSE}, all sequence variants
of each oligo will be presented.}
}
\value{
A \code{Biostrings::DNAStringSet} object
}
\description{
\code{convertToDNAStringSet()} can be used for converting all sequence
variants of all oligos or assays to a \code{Biostrings::DNAStringSet} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{RprimerOligo}: 

\item \code{RprimerAssay}: 
}}

\examples{
data("exampleRprimerOligo")
convertToDNAStringSet(exampleRprimerOligo[1:2, ])

convertToDNAStringSet(exampleRprimerOligo[1:2, ], asIUPAC = FALSE)

convertToDNAStringSet(exampleRprimerOligo[1:2, ], asRc = TRUE)
data("exampleRprimerAssay")
convertToDNAStringSet(exampleRprimerAssay[1:2, ], asRc = FALSE)

convertToDNAStringSet(exampleRprimerAssay[1:2, ], asIUPAC = FALSE)
}
\references{
H. Pages, P. Aboyoun, R. Gentleman and S. DebRoy (2020). Biostrings:
Efficient manipulation of biological strings. R package version
2.57.2.
}
