% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleRprimerAlignment}
\alias{exampleRprimerAlignment}
\title{An alignment of hepatitis E virus sequences}
\format{
Add here ############################################
}
\source{
The sequences were collected from NCBI GenBank,
}
\usage{
data(exampleRprimerAlignment)
}
\description{
An alignment with 621 hepatitis E virus sequences.
}
\keyword{datasets}
