% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleRprimerAssay}
\alias{exampleRprimerAssay}
\title{PCR assays for hepatitis E virus}
\format{
An \code{RprimerAssay} object.
}
\usage{
data("exampleRprimerAssay")
}
\description{
The object contains the following information:

\describe{
  \item{start}{Position where the assay starts.}
  \item{end}{Position where the assay ends.}
  \item{ampliconLength}{Length of the amplicon.}
  \item{tmDifferencePrimer}{Difference in Tm between
  the forward and reverse primer, absolute value.}
  \item{meanIdentity}{Average identity score of the primers
  (and probe if selected)}.
  \item{totalDegeneracy}{Total number of oligos in the assay.}
  \item{startFwd}{Position where the forward primer starts.}
  \item{endFwd}{Position where the reverse primer ends.}
  \item{lengthFwd}{Length of the forward primer.}
  \item{majorityFwd}{Majority sequence of the forward primer.}
  \item{gcMajorityFwd}{GC-content of the forward primer
  (majority sequence), proportion.}
  \item{identityFwd}{Average identity of the forward primer.}
  \item{tmMajorityFwd}{Tm of the forward primer
  (majority sequence).}
  \item{iupacFwd}{IUPAC sequence (i.e. with degenerate bases)
  of the forward primer.}
  \item{degeneracyFwd}{Number of variants of the forward primer}.
  \item{startRev}{Position where the reverse primer starts.}
  \item{endRev}{Position where the reverse primer ends.}
  \item{lengthRev}{Length of the reverse primer.}
  \item{majorityRev}{Majority sequence of the reverse primer.}
  \item{gcMajorityRev}{GC-content of the reverse primer
  (majority sequence), proportion.}
  \item{tmMajorityRev}{Tm of the reverse primer
  (majority sequence).}
  \item{identityRev}{Average identity of the reverse primer.}
  \item{iupacRev}{IUPAC sequence (i.e. with degenerate bases)
  of the reverse primer.}
  \item{degeneracyRev}{Number of variants of the reverse primer.}
  \item{allFwd}{Lists with all sequence variants of the forward primer.}
  \item{gcAllFwd}{Lists with the GC content of all
  sequence variants of the forward primer.}
  \item{tmAllFwd}{Lists with the Tm of all sequence variants of
  the reverse primer.}
  \item{allRev}{Lists with all sequence variants of the reverse primer.}
  \item{gcAllRev}{Lists with the GC content of all
  sequence variants of the forward primer.}
  \item{tmAllRev}{Lists with the Tm of all sequence variants of
  the reverse primer.}
}
}
\keyword{datasets}
