% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleRprimerAssay}
\alias{exampleRprimerAssay}
\title{PCR assays targeted to hepatitis E virus}
\format{
An \code{RprimerAssay} object.
}
\usage{
data("exampleRprimerAssay")
}
\description{
PCR assays targeted to hepatitis E virus
}
\keyword{datasets}
