% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleRprimerAssay}
\alias{exampleRprimerAssay}
\title{PCR assays targeted to hepatitis E virus}
\format{
An \code{RprimerAssay} object.
}
\usage{
data("exampleRprimerAssay")
}
\description{
The object contains the following information:

\describe{
  \item{start}{Position where the assay starts.}
  \item{end}{Position where the assay ends.}
  \item{ampliconLength}{Length of the amplicon.}
  \item{score}{Summarized oligo score.}
  \item{plusPr}{If the probe is valid in positive sense.}
  \item{startFwd}{Position where the forward primer starts.}
  \item{endFwd}{Position where the forward primer ends.}
  \item{lengthFwd}{Length of the forward primer.}
  \item{iupacSequenceFwd}{IUPAC sequence of the forward primer.}
  \item{coverageFwd}{Average coverage of the forward primer.}
  \item{degeneracyFwd}{Number of variants of the forward primer.}
  \item{gcContentMeanFwd}{Mean GC-content of the forward primer.}
  \item{gcContentRangeFwd}{Range in GC-content of the forward primer.}
  \item{tmMeanFwd}{Mean tm of the forward primer.}
  \item{tmRangeFwd}{Range in tm of the forward primer.}
  \item{sequenceFwd}{Sequence of the forward primer, all sequence variants.}
  \item{gcContentFwd}{GC-content of the forward primer,
  all sequence variants.}
  \item{tmFwd}{Tm of the forward primer, all sequence variants.}
  \item{dHFwd}{delta H of the forward
  primer (in cal/mol), all sequence variants.}
  \item{dSFwd}{delta S of
  the forward primer (in cal/K/mol), all sequence variants.}
  \item{startRev}{Position where the reverse primer starts.}
  \item{endRev}{Position where the reverse primer ends.}
  \item{lengthRev}{Length of the reverse primer.}
  \item{iupacSequenceRev}{IUPAC sequence of the reverse primer.}
  \item{coverageRev}{Average coverage of the reverse primer.}
  \item{degeneracyRev}{Number of variants of the reverse primer.}
  \item{gcContentMeanRev}{Mean GC-content of the reverse primer.}
  \item{gcContentRangeRev}{Range in GC-content of the reverse primer.}
  \item{tmMeanRev}{Mean tm of the reverse primer.}
  \item{tmRangeRev}{Range in tm of the reverse primer.}
  \item{sequenceRev}{Sequence of the reverse primer, all sequence variants.}
  \item{gcContentRev}{GC-content of the reverse primer, all sequence variants.}
  \item{tmRev}{Tm of the reverse primer, all sequence variants.}
  \item{dHRev}{delta H of the reverse
  primer (in cal/mol), of all sequence variants.}
  \item{dSRev}{delta S of the reverse primer (in cal/K/mol),
  all sequence variants.}
  \item{plusPr}{If the probe is valid in negative sense.}
  \item{minusPr}{If the probe is valid in negative sense.}
  \item{startPr}{Position where the probe starts.}
  \item{endPr}{Position where the probe ends.}
  \item{lengthPr}{Length of the probe.}
  \item{iupacSequencePr}{IUPAC sequence of the probe.}
  \item{coveragePr}{Average coverage of the probe.}
  \item{degeneracyPr}{Number of variants of the probe.}
  \item{gcContentMeanPr}{Mean GC-content of the probe.}
  \item{gcContentRangePr}{Range in GC-content of the probe.}
  \item{tmMeanPr}{Mean tm of the probe.}
  \item{tmRangePr}{Range in tm of the probe.}
  \item{sequencePr}{Sequence of the probe, all sequence variants.}
  \item{gcContentPr}{GC-content of the probe, all sequence variants.}
  \item{tmPr}{Tm of the probe, all sequence variants.}
  \item{dHPr}{delta H of the probe (in cal/mol), all sequence variants.}
  \item{dSPr}{delta S of the probe (in cal/K/mol), all sequence variants.}
  \item{roiStart}{Start position of the input consensus profile
  used for oligo design.}
  \item{roiEnd}{End position of the input consensus profile used
  for oligo design.}
}

Tm was calculated for perfectly matching oligo-target duplexes at a
sodium ion
concentration of 0.05 M, a primer
concentration of 500 nM and a probe concentration of 250 nM.

Delta H was calculated for perfectly matching oligo-target duplexes at a
sodium ion
concentration of 0.05 M.
}
\keyword{datasets}
