% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleRprimerMatchAssay}
\alias{exampleRprimerMatchAssay}
\title{Proportion of sequences that matches to assays}
\format{
An \code{RprimerMatchAssay} object.
}
\usage{
data("exampleRprimerMatchAssay")
}
\description{
The object contains the following information:

\describe{
  \item{iupacSequenceFwd}{The forward primer sequence in IUPAC format.}
  \item{perfectMatchFwd}{Proportion of target sequences that matches
  perfectly
  with the forward primer withing the intended binding region.}
  \item{idPerfectMatchFwd}{Names of all sequences that matches perfectly.}
  \item{oneMismatchFwd}{Proportion of target sequences with one mismatch to
  the forward primer
  within the intended binding region.}
  \item{idOneMismatchFwd}{Names of all sequences that matches with one
  mismatch.}
  \item{twoMismatchesFwd}{Proportion of target sequences with two mismatches
  to the forward primer
  within the intended binding region.}
  \item{idTwoMismatchesFwd}{Names of all sequences that matches with two
  mismatches.}
  \item{threeMismatchesFwd}{Proportion of target sequences with
  three mismatches to the forward primer
  within the intended binding region.}
  \item{idThreeMismatchesFwd}{Names of all sequences that matches with three
  mismatches.}
  \item{fourOrMoreMismatchesFwd}{Proportion of target sequences with
  four or more mismatches to the forward primer
  within the intended binding region.}
  \item{idFourOrMoreMismatchesFwd}{Names of all sequences that matches with
  four or more mismatches.}
  \item{offTargetMatchFwd}{Proportion of target sequences that matches
  to the forward primer with
  no more than four mismatches to all other regions within
  the target alignment.}
  \item{idOffTargetMatchFwd}{Names of all target sequences that matches
  to the forward primer with
  no more than four mismatches to all other regions within
  the target alignment.}
  \item{iupacSequenceRev}{The reverse primer sequence in IUPAC format.}
  \item{perfectMatchRev}{Proportion of target sequences that matches
  perfectly
  with the reverse primer withing the intended binding region.}
  \item{idPerfectMatchRev}{Names of all sequences that matches perfectly.}
  \item{oneMismatchRev}{Proportion of target sequences with one mismatch to
  the reverse primer
  within the intended binding region.}
  \item{idOneMismatchRev}{Names of all sequences that matches with one
  mismatch.}
  \item{twoMismatchesRev}{Proportion of target sequences with two mismatches
  to the reverse primer
  within the intended binding region.}
  \item{idTwoMismatchesRev}{Names of all sequences that matches with two
  mismatches.}
  \item{threeMismatchesRev}{Proportion of target sequences with
  three mismatches to the reverse primer
  within the intended binding region.}
  \item{idThreeMismatchesRev}{Names of all sequences that matches with three
  mismatches.}
  \item{fourOrMoreMismatchesRev}{Proportion of target sequences with
  four or more mismatches to the reverse primer
  within the intended binding region.}
  \item{idFourOrMoreMismatchesRev}{Names of all sequences that matches with
  four or more mismatches.}
  \item{offTargetMatchRev}{Proportion of target sequences that matches
  to the reverse primer with
  no more than four mismatches to all other regions within
  the target alignment.}
  \item{idOffTargetMatchRev}{target sequences that matches
  to the reverse primer with
  no more than four mismatches to all other regions within
  the target alignment.}
  \item{iupacSequencePr}{The probe sequence in IUPAC format.}
  \item{perfectMatchPr}{Proportion of target sequences that matches perfectly
  with the probe withing the intended binding region.}
  \item{idPerfectMatchPr}{Names of all sequences that matches perfectly.}
  \item{oneMismatchPr}{Proportion of target sequences with one mismatch to
  the probe
  within the intended binding region.}
  \item{idOneMismatchPr}{Names of all sequences that matches with one
  mismatch.}
  \item{twoMismatchesPr}{Proportion of target sequences with two mismatches
  to the probe
  within the intended binding region.}
  \item{idTwoMismatchesPr}{Names of all sequences that matches with two
  mismatches.}
  \item{threeMismatchesPr}{Proportion of target sequences with
  three mismatches to the probe
  within the intended binding region.}
  \item{idThreeMismatchesPr}{Names of all sequences that matches with three
  mismatches.}
  \item{fourOrMoreMismatchesPr}{Proportion of target sequences with
  four or more mismatches to the probe
  within the intended binding region.}
  \item{idFourOrMoreMismatchesPr}{Names of all sequences that matches with
  four or more mismatches.}
  \item{offTargetMatchPr}{Proportion of target sequences that matches
  to the probe with
  no more than four mismatches to all other regions within
  the target alignment.}
  \item{idOffTargetMatchPr}{Names of all target sequences that matches
  to the probe with
  no more than four mismatches to all other regions within
  the target alignment.}
 }
}
\details{
This dataset describes the proportion of HEV target sequences that
matches to a set of assays.
}
\keyword{datasets}
