% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleRprimerMatchOligo}
\alias{exampleRprimerMatchOligo}
\title{Proportion of sequences that matches to oligos}
\format{
An \code{RprimerMatchOligo} object.
}
\usage{
data("exampleRprimerMatchOligo")
}
\description{
This dataset describes the proportion of HEV target sequences that
matches to a set of oligos.
}
\keyword{datasets}
