% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleRprimerMatchOligo}
\alias{exampleRprimerMatchOligo}
\title{Proportion of matching sequences to a target}
\format{
An \code{RprimerMatchOligo} object.
}
\usage{
data("exampleRprimerMatchOligo")
}
\description{
The object contains the following information:

\describe{
  \item{iupacSequence}{The oligo sequence in IUPAC format.}
  \item{perfectMatch}{Proportion of target sequences that matches perfectly
  to the oligo within the intended oligo binding region.}
  \item{oneMismatch}{Proportion of target sequences with one mismatch to
  the oligo
  within the intended oligo binding region.}
  \item{twoMismatches}{Proportion of target sequences with two mismatches
  to the oligo
  within the intended oligo binding region.}
  \item{threeMismatches}{Proportion of target sequences with
  three mismatches to the oligo
  within the intended oligo binding region.}
  \item{fourOrMoreMismatches}{Proportion of target sequences with
  four or more mismatches to the oligo
  within the intended oligo binding region.}
  \item{offTargetMatch}{Proportion of target sequences that matches
  to the oligo with
  no more than four mismatches to all other regions of the target alignment.}
 }
}
\details{
This dataset describes the proportion of oligo sequences that
matches to a set of HEV target sequences.
}
\keyword{datasets}
