% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleRprimerMatchOligo}
\alias{exampleRprimerMatchOligo}
\title{Proportion of matching sequences to a target}
\format{
An \code{RprimerMatchOligo} object.
}
\usage{
data("exampleRprimerMatchOligo")
}
\description{
The object contains the following information:

\describe{
  \item{iupacSequence}{The oligo sequence in IUPAC format}
  \item{n0mm}{Proportion of target sequences with zero mismatches.}
  \item{n1mm}{Proportion of target sequences with one mismatch.}
  \item{n2mm}{Proportion of target sequences with two mismatches.}
  \item{n3mm}{Proportion of target sequences with three mismatches.}
  \item{n4orMoremm}{Proportion of target sequences with four or more
  mismatches.}
 }
}
\details{
This dataset describes the proportion of oligo sequences that
matches to a set of HEV target sequences.
}
\keyword{datasets}
