% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleRprimerOligo}
\alias{exampleRprimerOligo}
\title{Primers and probes targeted to hepatitis E virus}
\format{
An \code{RprimerOligo} object.
}
\usage{
data("exampleRprimerOligo")
}
\description{
Primers and probes targeted to hepatitis E virus
}
\keyword{datasets}
