% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleRprimerOligo}
\alias{exampleRprimerOligo}
\title{Oligonucleotides for hepatitis E virus}
\format{
An \code{RprimerOligo} object.
}
\usage{
data("exampleRprimerOligo")
}
\description{
The object contains the following information:

\describe{
  \item{type}{Type of oligo, primer or probe.}
  \item{start}{Position where the oligo starts.}
  \item{end}{Position where the oligo ends.}
  \item{length}{Length of the oligo.}
  \item{majority}{Majority sequence.}
  \item{majorityRc}{Majority sequence, reverse complement.}
  \item{gcMajority}{GC content (majority sequence), proportion.}
  \item{tmMajority}{Melting temperature.}
  \item{identity}{Average identity score of the oligo.}
  \item{iupac}{IUPAC sequence (i.e. with degenerate bases).}
  \item{iupacRc}{IUPAC sequence, reverse complement.}
  \item{degeneracy}{Number of variants of the degenerate oligo.}
  \item{all}{Lists with all sequence variants of the oligos.}
  \item{allRc}{Lists with all sequence variants of the oligos, reverse
  complements.}
  \item{gcAll}{Lists with the GC content of all
  sequence variants of the oligos.}
  \item{tmAll}{Lists with the Tm of all sequence variants of the oligos.}
  \item{alignmentStart}{Start position of the input consensus profile
  used for oligo design.}
  \item{alingnmentEnd}{End position of the input consensus profile used
  for oligo design.}
}
}
\keyword{datasets}
