% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleRprimerProfile}
\alias{exampleRprimerProfile}
\title{A consensus profile from an alignment of hepatitis E virus sequences}
\format{
An \code{RprimerProfile} object.
}
\usage{
data("exampleRprimerProfile")
}
\description{
A consensus profile from an alignment of hepatitis E virus sequences
}
\keyword{datasets}
