% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleRprimerProfile}
\alias{exampleRprimerProfile}
\title{Consensus profile of an alignment of hepatitis E virus sequences}
\format{
An \code{RprimerProfile} object.
}
\source{
The sequences were collected from NCBI GenBank.
}
\usage{
data("exampleRprimerProfile")
}
\description{
Consensus profile of an alignment of hepatitis E virus sequences
}
\keyword{datasets}
