% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleRprimerProperties}
\alias{exampleRprimerProperties}
\title{Sequence properties of hepatitis E virus sequences}
\format{
An \code{RprimerProperties} object.

\describe{
  \item{Position}{Position in the alignment.}
  \item{Majority}{Majority consensus sequence.}
  \item{IUPAC}{IUPAC consensus sequence.}
  \item{Gaps}{Proportion of gaps.}
  \item{Identity}{Proportion of the most common nucleotide.}
  \item{Entropy}{Shannon entropy}
}
}
\source{
The sequences were collected from NCBI GenBank.
}
\usage{
data("exampleRprimerProperties")
}
\description{
Sequence properties of hepatitis E virus sequences
}
\keyword{datasets}
