% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{get_assays}
\alias{get_assays}
\title{Get (RT)-PCR assays from oligos}
\usage{
get_assays(x, length = 65:120, max_tm_difference = 1)
}
\arguments{
\item{x}{an object of class "rprimer_oligo".}

\item{length}{Desired amplicon length,
ranging from 40 to 5000 base pairs. The
default is \code{65:120}, which means that assays with amplicon lengths from
65 to 120 base-pairs will be considered as acceptable.}

\item{max_tm_difference}{Maximum Tm difference (in C) between the two primers
(absolute value). A number between 0 and 30. The default is 1.}
}
\value{
A tibble (a data frame) of class 'rprimer_assay' with all candidate
assays.
An error message will return if no assays are found.
}
\description{
\code{get_assays} combines forward and reverse primers to (RT)-PCR assays.
}
\details{
#Warning:
The Tm-difference is calculated from the majority oligos, and
may thus be misleading for degenerate (iupac) oligos.
}
