% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignmentProfile.R
\name{getAlignmentProfile}
\alias{getAlignmentProfile}
\title{Get the sequence profile of an alignment}
\usage{
getAlignmentProfile(x)
}
\arguments{
\item{x}{A Biostrings::DNAMultipleAlignment object.}
}
\value{
An RprimerProfile object, containing a numeric matrix with the
proportion of each
nucleotide at each position within an alignment
of DNA sequences. The matrix has six rows,
named 'A', 'C', 'G', 'T', '-' and 'Other'. '-' represents gaps and
'Other' represents nucleotides other than A, C, G and T, e.g.
wobble bases. The columns are named
according to which position they correspond to in the input alignment.
}
\description{
\code{getAlignmentProfile} returns a matrix with the
proportion of each nucleotide at each position within an alignment
of DNA sequences.The function is a wrapper around
Biostrings::consensusMatrix.
}
\examples{
data("exampleRprimerAlignment")
getAlignmentProfile(exampleRprimerAlignment)

}
