% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssays.R
\name{getAssays}
\alias{getAssays}
\title{Get (RT)-PCR assays from oligos}
\usage{
getAssays(
  primers,
  probes = NULL,
  length = 65:120,
  maxTmDifferencePrimers = 2,
  tmDifferenceProbes = c(0, 20)
)
}
\arguments{
\item{primers}{An \code{RprimerOligo} object.}

\item{probes}{Optional, defaults to \code{NULL}.
An \code{RprimerOligo} object if probes are to be used.}

\item{length}{Amplicon length, a numeric vector [40, 5000]. Defaults to
\code{65:120}.}

\item{maxTmDifferencePrimers}{Maximum Tm difference between the two primers
(absolute value). A number [0, 30]. Defaults to 2.
Note that the Tm-difference is calculated from the majority primers, and
may thus be misleading for degenerate (IUPAC) primers.}

\item{tmDifferenceProbes}{Acceptable Tm difference between the primers (average Tm of the
primer pair) and probe. A numeric vector [-20, 20],
defaults to \code{c(0, 20)}.
The Tm-difference is calculated by subtracting the
Tm of the probe with the average Tm of the majority
primer pair. Thus, a negative Tm-difference
means that the Tm of the probe is lower than the average Tm of the
primer pair.
Note that the Tm-difference is calculated from the majority oligos, and
may thus be misleading for degenerate (IUPAC) oligos.}
}
\value{

}
\description{
\code{getAssays()} combines forward and reverse primers
and (if selected) probes to (RT)-PCR assays.
}
\seealso{
getOligos
}
