% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOligos.R
\name{getOligos}
\alias{getOligos}
\title{Get oligos from sequence properties}
\usage{
getOligos(
  x,
  target,
  length = 18:22,
  maxGapFrequency = 0.1,
  maxDegeneracy = 4,
  avoid_3end_ta = TRUE,
  avoid_3end_runs = TRUE,
  avoid_gc_rich_3end = TRUE,
  avoid_5end_g = FALSE,
  gcRange = c(0.45, 0.55),
  tmRange = c(48, 70),
  concOligo = 5e-07,
  concNa = 0.05
)
}
\arguments{
\item{x}{an rprimerProperties object.}

\item{length}{Oligo length. A vector of integers. The minimum allowed
value is 14 and the maximum allowed value is 30.
It defaults to \code{18:22}.}

\item{maxGapFrequency}{Maximum allowed gap frequency.
A number [0, 1). It defaults to 0.1, meaning that
positions with a gap frequency above 0.1 will not be
considered as oligo regions.}

\item{maxDegeneracy}{Maximum number of variants of each oligo.
The minimum allowed value is 1 and the maximum
allowed value is 16. The default is 4.}

\item{avoid_3end_ta}{\code{TRUE} or \code{FALSE}.
If oligos with T or A at the 3' end
should be replaced with \code{NA}
(recommended for primers). The default is \code{TRUE}.}

\item{avoid_3end_runs}{\code{TRUE} or \code{FALSE}.
If oligos with more than two runs
of the same nucleotide at the 3' end should be replaced with \code{NA}
(recommended for primers).
The default is \code{TRUE}.}

\item{avoid_gc_rich_3end}{\code{TRUE} or \code{FALSE}.
If oligos with more than three G or C within the last five bases of the 3'
end should be replaced with \code{NA}
(recommended for primers).
The default is \code{TRUE}.}

\item{avoid_5end_g}{\code{TRUE} or \code{FALSE}.If oligos with g
at the 5' end should be replaced with \code{NA}
(recommended for probes). The default is \code{FALSE}.}

\item{gcRange}{Oligo GC-content range (proportion) Can range between
0 and 1. The default is \code{c(0.45, 0.55)}.}

\item{tmRange}{The Tm-range of each oligo. Can range between 20 and 90.
The default is \code{c(48, 70)}.}

\item{concOligo}{The concentration of oligonucleotide in M,
ranging from 0.2e-07 M (20 nM) to 2e-06 M (2000 nM).
The default value is 5e-07 M (500 nM) (for Tm calculation)}

\item{concNa}{The sodium ion concentration in M, ranging
from 0.01 M to 1 M. The default value is 0.05 M (50 mM)
(for Tm calculation).}
}
\value{
A tibble (a data frame) of class 'rprimer_oligo', with all oligo
candidates. An error message will return if no oligos are found.

The tibble contains the following information:

\describe{
  \item{begin}{position where the oligo begins}
  \item{end}{position where the oligo ends}
  \item{length}{length of the oligo}
  \item{majority}{majority sequence}
  \item{iupac}{iupac sequence (i.e. with degenerate bases)}
  \item{majority_rc}{majority sequence, reverse complement}
  \item{iupac_rc}{iupac sequence, reverse complement}
  \item{degenerates}{number of degenerate bases}
  \item{degeneracy}{number of variants}
  \item{gc_majority}{gc-content (majority sequence), proportion}
  \item{tm_majority}{melting temperature (majority sequence),
  degrees Celcius}
  \item{pm_majority}{proportion of perfectly matching sequences,
  (to the target alignment) majority sequence}
  \item{pm_iupac}{proportion of perfectly matching sequences,
  iupac sequence}
  \item{match_matrix}{a logical matrix describing which sequences
  in the target alignment the oligo matches perfectly to}
}
}
\description{
\code{getOligos} identifies oligos (primers and probes) from
sequence properties.
}
\section{Excluded oligos}{

The function excludes oligos with:
* More than than three consecutive runs of the same dinucleotide
(e.g. 'tatatata')
* More than four consecutive runs of the
same nucleotide
* Oligos that are duplicated
}

\section{Tm}{

The melting temperature is calculated using the nearest-neighbour method.

* Oligos are not expected to be self-complementary, so no symmetry
correction is done
* The oligo concentration is assumed to be much higher
than the target concentration

See references for table values and equations.
}

\section{Note}{

GC-content and Tm are calculated based on the majority oligos, and
may thus be misleading for degenerate (iupac) oligos.
}

\examples{
data("exampleRprimerProperties")

get_oligos(
exampleRprimerProperties,
length = 18:22,
maxGapFrequency = 0.1,
maxDegeneracy = 4,
avoid_3end_ta = TRUE,
avoid_3end_runs = TRUE,
avoid_gc_rich_3end = TRUE,
avoid_5end_g = FALSE,
gc_range = c(0.45, 0.55),
tm_range = c(48, 70)
)

}
\references{
# Tm-calculation:

SantaLucia, J, et al. (1996)
Improved Nearest-Neighbor Parameters for Predicting DNA Duplex Stability.
Biochemistry, 35: 3555-3562 (Formula and salt correction method
are from here)

Allawi, H. & SantaLucia, J. (1997)
Thermodynamics and NMR of Internal G-T Mismatches in DNA.
Biochemistry, 36, 34: 10581–10594
(Duplex initiation parameters are from here)

SantaLucia, J (1998) A unified view of polymer,
dumbell, and oligonucleotide DNA nearest-neighbor thermodynamics.
Proc. Natl. Acad. Sci. USA, 95: 1460-1465. (Table values are from here)
}
