% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOligos.R
\name{getOligos}
\alias{getOligos}
\title{Get oligos}
\usage{
getOligos(
  x,
  length = 18:22,
  maxGapFrequency = 0.1,
  maxDegeneracy = 4,
  gcClamp = TRUE,
  avoid3EndRuns = TRUE,
  avoid5EndG = FALSE,
  gcRange = c(0.45, 0.55),
  tmRange = c(48, 70),
  concOligo = 5e-07,
  concNa = 0.05,
  showAllVariants = TRUE
)
}
\arguments{
\item{x}{an RprimerProperties object.}

\item{length}{Oligo length. The minimum allowed
value is 14 and the maximum allowed value is 30.
It defaults to \code{18:22}.}

\item{maxGapFrequency}{Maximum allowed gap frequency.
A number [0, 1]. It defaults to 0.1, which means that
positions with a gap frequency above 0.1 will not be
considered as an oligo region.}

\item{maxDegeneracy}{Maximum number of variants.
The minimum allowed value is 1 and the maximum
allowed value is 32. It defaults to 4.}

\item{gcClamp}{\code{TRUE} or \code{FALSE}.
If oligos with no GC-clamp
should be replaced with \code{NA}
(recommended for primers). It defaults to \code{TRUE}.}

\item{avoid3EndRuns}{\code{TRUE} or \code{FALSE}.
If oligos with more than two runs
of the same nucleotide at the 3' end should be replaced with \code{NA}
(recommended for primers).
It defaults to \code{TRUE}.}

\item{avoid5EndG}{\code{TRUE} or \code{FALSE}.If oligos with g
at the 5' end should be replaced with \code{NA}
(recommended for probes). It defaults to \code{FALSE}.}

\item{gcRange}{Accepted GC-content (proportion, not %). A numeric vector [0, 1].
It defaults to \code{c(0.45, 0.55)}.}

\item{tmRange}{Accepted Tm. A numeric vector [30, 90].
It defaults to \code{c(48, 70)}.}

\item{concOligo}{Oligo concentration in M, for Tm calculation. A numeric vector
[0.2e-07, 2e-06], i.e. between 20 nM and 20000 nM.
It defaults to 5e-07 M (500 nM).}

\item{concNa}{The sodium ion concentration in M, for Tm calculation.
A numeric vector [0.01, 1]. It defaults to 0.05 M (50 mM).}

\item{showAllVariants}{If sequence, GC-content and Tm should be presented for all
variants of each oligo (in case of degenerate bases).
\code{TRUE} or \code{FALSE}. It defaults to \code{TRUE}.}
}
\value{
A tibble (a data frame) with all oligo
candidates. An error message will return if no oligos are found.

The tibble contains the following information:

\describe{
  \item{Begin}{position where the oligo begins}
  \item{End}{position where the oligo ends}
  \item{Length}{length of the oligo}
  \item{Majority}{majority sequence}
  \item{IUPAC}{IUPAC sequence (i.e. with degenerate bases)}
  \item{Majority_RC}{majority sequence, reverse complement}
  \item{IUPAC_RC}{IUPAC sequence, reverse complement}
  \item{Degeneracy}{number of variants}
  \item{GC_majority}{GC content (majority sequence), proportion}
  \item{Tm_majority}{melting temperature}
}
}
\description{
\code{getOligos} identifies oligos (primers and probes) from
sequence properties.
}
\section{Excluded oligos}{

The function excludes all oligos with:
* More than than three consecutive runs of the same dinucleotide
(e.g. 'TATATATA')
* More than four consecutive runs of the
same nucleotide (e.g. 'AAAAA')
It also excludes oligos that are duplicated, to prevent binding to several
places in the target.
These checks are made on the majority oligo sequences.
}

\section{Tm}{

The melting temperature is calculated using the nearest-neighbor method,
with the following assumptions:

* Oligos are not expected to be self-complementary (hence no symmetry
correction is done)
* The oligo concentration is assumed to be much higher
than the target concentration

See references for table values and equations.
}

\section{Note}{

GC-content and Tm are calculated based on the majority oligos, and
may thus be misleading for degenerate (IUPAC) oligos.
}

\examples{
data("exampleRprimerProperties")

getOligos <- function(
exampleRprimerProperties,
length = 18:22,
maxGapFrequency = 0.1,
maxDegeneracy = 4,
gcClamp = TRUE,
avoid3EndRuns = TRUE,
avoid5EndG = FALSE,
gcRange = c(0.45, 0.55),
tmRange = c(48, 65),
concOligo = 5e-07,
concNa = 0.05,
showAllVariants = TRUE
)

}
\references{
Tm-calculation:

SantaLucia, J, et al. (1996)
Improved Nearest-Neighbor Parameters for Predicting DNA Duplex Stability.
Biochemistry, 35: 3555-3562 (Formula and salt correction method
are from here)

Allawi, H. & SantaLucia, J. (1997)
Thermodynamics and NMR of Internal G-T Mismatches in DNA.
Biochemistry, 36, 34: 10581–10594
(Duplex initiation parameters are from here)

SantaLucia, J (1998) A unified view of polymer,
dumbell, and oligonucleotide DNA nearest-neighbor thermodynamics.
Proc. Natl. Acad. Sci. USA, 95: 1460-1465. (Table values are from here)
}
