% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOligos.R
\name{getOligos}
\alias{getOligos}
\title{Get oligos}
\usage{
getOligos(
  x,
  lengthPrimer = 18:22,
  maxGapFrequencyPrimer = 0.1,
  maxDegeneracyPrimer = 4,
  gcClampPrimer = TRUE,
  avoid3EndRunsPrimer = TRUE,
  minEndIdentityPrimer = 0.98,
  gcRangePrimer = c(0.45, 0.55),
  tmRangePrimer = c(55, 65),
  concPrimer = 5e-07,
  probe = TRUE,
  lengthProbe = 18:22,
  maxGapFrequencyProbe = 0.1,
  maxDegeneracyProbe = 4,
  avoid5EndGProbe = TRUE,
  gcRangeProbe = c(0.45, 0.55),
  tmRangeProbe = c(55, 70),
  concProbe = 2.5e-07,
  concNa = 0.05,
  showAllVariants = TRUE
)
}
\arguments{
\item{x}{An \code{RprimerProfile} object.}

\item{lengthPrimer}{Primer length. A numeric vector [14, 30].
Defaults to \code{18:22}.}

\item{maxGapFrequencyPrimer}{Maximum allowed gap frequency for primers.
A number [0, 1]. Defaults to 0.1.}

\item{maxDegeneracyPrimer}{Maximum number of variants of each primer. A number [1, 32]. Defaults to 4.}

\item{gcClampPrimer}{\code{TRUE} or \code{FALSE}.
If primers with no GC-clamp
should be replaced with \code{NA}.
Defaults to \code{TRUE}. A GC-clamp
is identified as two to three G or
C:s within the last five bases (3' end) of the oligo.}

\item{avoid3EndRunsPrimer}{\code{TRUE} or \code{FALSE}.
If primers with more than two runs
of the same nucleotide at the 3' end should be replaced with \code{NA}.
Defaults to \code{TRUE}.}

\item{minEndIdentityPrimer}{Optional. If specified, a number [0, 1]. The minimum allowed identity
at the 3' end of the primer (i.e. the last five bases).}

\item{gcRangePrimer}{GC-content range for primers (proportion, not %). A numeric vector [0, 1].
Defaults to \code{c(0.45, 0.55)}.}

\item{tmRangePrimer}{Tm range for primers.
A numeric vector [30, 90]. Defaults to \code{c(55, 65)}.
Tm is calculated using the nearest-neighbor method,
with the following assumptions:
1) Oligos are not expected to be self-complementary (i.e. no symmetry
correction is done);
2) The oligo concentration is assumed to be much higher
than the target concentration. See references for table values and equations.}

\item{concPrimer}{Primer concentration in M, for Tm calculation. A numeric vector
[0.2e-07, 2e-06], i.e. between 20 nM and 20000 nM.
It defaults to 5e-07 M (500 nM).}

\item{probe}{If probes should be designed as  well. \code{TRUE} or \code{FALSE},
defaults to \code{TRUE}.}

\item{lengthProbe}{Probe length. A numeric vector [14, 30].
Defaults to \code{18:22}.}

\item{maxGapFrequencyProbe}{Maximum allowed gap frequency for probes.
A number [0, 1]. Defaults to 0.1.}

\item{maxDegeneracyProbe}{Maximum number of variants of each probe. A number [1, 32]. Defaults to 4.}

\item{avoid5EndGProbe}{\code{TRUE} or \code{FALSE}. If probes with G
at the 5' end should be excluded. Defaults to \code{TRUE}.}

\item{gcRangeProbe}{GC-content range for probes (proportion, not %). A numeric vector [0, 1].
Defaults to \code{c(0.45, 0.55)}.}

\item{tmRangeProbe}{Tm range for probes.
A numeric vector [30, 90]. Defaults to \code{c(55, 65)}.
Tm is calculated using the nearest-neighbor method,
with the following assumptions:
1) Oligos are not expected to be self-complementary (i.e. no symmetry
correction is done);
2) The oligo concentration is assumed to be much higher
than the target concentration. See references for table values and equations.}

\item{concProbe}{Probe concentration in M, for Tm calculation. A numeric vector
[0.2e-07, 2e-06], i.e. between 20 nM and 20000 nM.
It defaults to 5e-07 M (500 nM).}

\item{concNa}{The sodium ion concentration in the PCR reaction in M, for Tm calculation.
A numeric vector [0.01, 1]. It defaults to 0.05 M (50 mM).}

\item{showAllVariants}{If sequence, GC-content and Tm should be presented for all
variants of each oligo, not just for majority variants
(in case of degenerate bases).
\code{TRUE} (slower) or \code{FALSE} (faster).
Defaults to \code{TRUE}.

\code{getOligos()} excludes all oligos with
more than than three consecutive runs of the same dinucleotide
(e.g. 'TATATATA') and more than four consecutive runs of the
same nucleotide (e.g. 'AAAAA').
It also excludes oligos that are duplicated to prevent binding to several
places in the target. These checks are done on majority oligos.}
}
\value{
An \code{RprimerOligo} object.
An error message will return if no oligos are found.

The object contains the following information:

\describe{
  \item{type}{Type of oligo, primer or probe.}
  \item{start}{Position where the oligo starts.}
  \item{end}{Position where the oligo ends.}
  \item{length}{Length of the oligo.}
  \item{majority}{Majority sequence.}
  \item{majorityRc}{Majority sequence, reverse complement.}
  \item{gcMajority}{GC content (majority sequence), proportion.}
  \item{tmMajority}{Melting temperature.}
  \item{identity}{Average identity score of the oligo.}
  \item{iupac}{IUPAC sequence (i.e. with degenerate bases).}
  \item{iupacRc}{IUPAC sequence, reverse complement.}
  \item{degeneracy}{Number of variants of the degenerate oligo.}
  \item{alignmentStart}{Start position of the input consensus profile.}
  \item{alingnmentEnd}{End position of the input consensus profile.}
}

If \code{showAllVariants == TRUE}, the following columns are also included:

\describe{
  \item{all}{Lists with all sequence variants of the oligos.}
  \item{allRc}{Lists with all sequence variants of the oligos, reverse
  complements.}
  \item{gcAll}{Lists with the GC content of all
  sequence variants of the oligos.}
  \item{tmAll}{Lists with the Tm of all sequence variants of the oligos.}
}
}
\description{
\code{getOligos()} identifies oligos (primers and probes)
from sequence properties.
}
\examples{
data("exampleRprimerProfile")
## Design primers and probes with default values
getOligos(exampleRprimerProfile)

}
\references{
Tm-calculation:

Formula and salt correction method:
SantaLucia, J, et al. (1996)
Improved Nearest-Neighbor Parameters for Predicting DNA Duplex Stability.
Biochemistry, 35: 3555-3562

Duplex initiation parameters:
Allawi, H. & SantaLucia, J. (1997)
Thermodynamics and NMR of Internal G-T Mismatches in DNA.
Biochemistry, 36, 34: 10581–10594

Table values for nearest-neighbors:
SantaLucia, J (1998) A unified view of polymer,
dumbell, and oligonucleotide DNA nearest-neighbor thermodynamics.
Proc. Natl. Acad. Sci. USA, 95: 1460-1465.
}
