% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignmentProperties.R
\name{getSequenceProperties}
\alias{getSequenceProperties}
\title{Get alignment properties}
\usage{
getSequenceProperties(x, iupacThreshold = 0)
}
\arguments{
\item{x}{An RprimerProfile object.}

\item{iupacThreshold}{A number (0, 0.2].
At each position, all nucleotides with a proportion
higher or equal to the iupacThreshold will be included in
the IUPAC consensus sequence. The default is 0.}
}
\value{
An RprimerProperties object, which contains a tibble (a data frame)
with information on majority and IUPAC consensus sequence, gap frequency,
nucleotide identity and Shannon entropy.
}
\description{
\code{getAlignmentProperties} returns information from
an RprimerProfile
}
\section{Majority consensus sequence}{

The most frequently occurring nucleotide.
If two or more bases occur with the same frequency,
the consensus nucleotide will be randomly selected among these bases.
}

\section{IUPAC consensus sequence}{

The consensus sequence expressed in IUPAC format (i.e. with wobble bases)
Note that the IUPAC consensus sequence only
takes 'A', 'C', 'G', 'T' and '-' as input. Degenerate bases
present in the alignment will be skipped. If a position only contains
degenerate/invalid bases, the IUPAC consensus will be \code{NA} at that
position.
}

\section{Gaps}{

Proportion of gaps. Gaps are recognized as "-".
}

\section{Identity}{

Proportion of
the most common base. Gaps (-),
as well as bases other than A, C, G and T are excluded from the
calculation.
}

\section{Entropy}{

Shannon entropy is a measurement of
variability. First, for each nucleotide that occurs at a specific position,
\code{p*log2(p)}, is calculated, where \code{p} is the proportion of
that nucleotide. Then, these values are summarized,
and multiplied by \code{-1}.
A value of \code{0} indicate no variability and a high value
indicate high variability.
Gaps (-), as well as bases other than
A, C, G and T are excluded from the calculation.
}

\examples{
getAlignmentProperties(exampleRprimerProfile)
getAlignmentProperties(exampleRprimerProfile, iupacThreshold = 0.1)

}
