% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oligos.R
\name{oligos}
\alias{oligos}
\title{Design primers and probes}
\usage{
oligos(
  x,
  maxGapFrequency = 0.01,
  lengthPrimer = 18:22,
  maxDegeneracyPrimer = 4,
  gcClampPrimer = TRUE,
  avoidThreeEndRunsPrimer = TRUE,
  minThreeEndCoveragePrimer = 0.98,
  gcRangePrimer = c(0.4, 0.65),
  tmRangePrimer = c(50, 65),
  concPrimer = 500,
  designStrategyPrimer = "ambiguous",
  probe = TRUE,
  lengthProbe = 18:22,
  maxDegeneracyProbe = 4,
  avoidFiveEndGProbe = TRUE,
  gcRangeProbe = c(0.4, 0.65),
  tmRangeProbe = c(50, 70),
  concProbe = 250,
  concNa = 0.05,
  temperature = 60
)
}
\arguments{
\item{x}{An \code{RprimerProfile} object.}

\item{maxGapFrequency}{Maximum allowed gap frequency (for both primers and probes).
A number [0, 1], defaults to \code{0.01}.}

\item{lengthPrimer}{Primer length. A numeric vector [15, 30],
defaults to \code{18:22}.}

\item{maxDegeneracyPrimer}{Maximum number of variants of each primer. A number [1, 64],
defaults to \code{4}.}

\item{gcClampPrimer}{If primers must have a GC-clamp.
A GC-clamp
is identified as two to three G or
C:s within the last five bases (3' end) of the oligo.
\code{TRUE} or \code{FALSE}, defaults to \code{TRUE}.}

\item{avoidThreeEndRunsPrimer}{If primers with more than two runs
of the same nucleotide at the terminal 3' end should be avoided.
\code{TRUE} or \code{FALSE}, defaults to \code{TRUE}.}

\item{minThreeEndCoveragePrimer}{Minimum allowed coverage at the 3' end (the last five bases).
A number [0, 1]. A value of 1 means that all bases within the 3' end
must cover all sequence
variants in the target alignment. Defaults to \code{0.98}.}

\item{gcRangePrimer}{GC-content range for primers.
A numeric vector [0, 1], defaults to \code{c(0.40, 0.65)}.}

\item{tmRangePrimer}{Tm range for primers.
A numeric vector [30, 90], defaults to \code{c(55, 65)}.}

\item{concPrimer}{Primer concentration in nM, for Tm calculation. A number
[20, 2000], defaults to \code{500}.}

\item{designStrategyPrimer}{\code{"ambiguous"} or \code{"mixed"}. Defaults to \code{"ambiguous"}.}

\item{probe}{If probes should be designed. \code{TRUE} or \code{FALSE},
defaults to \code{TRUE}.}

\item{lengthProbe}{Probe length. A numeric vector [15, 40],
defaults to \code{18:22}.}

\item{maxDegeneracyProbe}{Maximum number of variants of each probe. A number [1, 64],
defaults to \code{4}.}

\item{avoidFiveEndGProbe}{If probes with G
at the 5' end should be avoided. \code{TRUE} or \code{FALSE},
defaults to \code{TRUE}.}

\item{gcRangeProbe}{GC-content range for probes (proportion). A numeric vector [0, 1],
defaults to \code{c(0.40, 0.65)}.}

\item{tmRangeProbe}{Tm range for probes.
A numeric vector [30, 90], defaults to \code{c(55, 70)}.}

\item{concProbe}{Primer concentration in nM, for tm calculation. A numeric vector
[20, 2000], defaults to \code{250}.}

\item{concNa}{The sodium ion concentration in the PCR reaction in M, for calculation of
tm and delta G.
A numeric vector [0.01, 1], defaults to \code{0.05} (50 mM).}

\item{temperature}{Annealing temperature in Celcius, for calculation of delta G.
A number [37, 70], defaults to \code{60}.}
}
\value{
An \code{RprimerOligo} object. An error message will return
if no oligos are found.
}
\description{
\code{oligos()} designs oligos (primers and probes)
from an \code{RprimerProfile} object.
}
\section{Output}{


The output contains the following information:

\describe{
  \item{type}{Whether the oligo is a primer or probe.}
  \item{fwd}{\code{TRUE} if the oligo is valid in forward direction,
    \code{FALSE} otherwise.}
  \item{rev}{\code{TRUE} if the oligo is valid in reverse direction,
    \code{FALSE} otherwise.}
  \item{start}{Start position of the oligo.}
  \item{end}{End positon of the oligo.}
  \item{length}{Oligo length.}
  \item{iupacSequence}{Oligo sequence, with amiguous bases (if any).}
  \item{iupaSequenceRc}{The reverse complement of the iupacSequence.}
  \item{identity}{For ambiguous oligos: Average identity of the oligo.
    For mixed oligos: Average identity of the 5' (consensus) part of the
    oligo. The value can range from 0 to 1.}
  \item{coverage}{For ambiguous oligos: Average coverage of the oligo.
    For mixed oligos: Average coverage of the 3' (degenerate) part of the
    oligo. The value can range from 0 to 1.}
  \item{degeneracy}{Number of sequence variants of the oligo.}
  \item{gcContentMean}{Mean GC-content of all sequence variants of the oligo.
  }
  \item{gcContentRange}{Range in GC-content of all sequence variants of
    the oligo.}
  \item{tmMean}{Mean tm of all sequence variants of the oligo.}
  \item{tmRange}{Range in tm of all sequence variants of the oligo.}
  \item{deltaGMean}{Mean delta G (in kcal/mole)
  of all sequence variants of the oligo.}
  \item{deltaGRange}{Range in delta G (in kcal/mole)
  of all sequence variants of the oligo.}
  \item{sequence}{All sequence variants of the oligo.}
  \item{sequenceRc}{Reverse complements of all sequence variants.}
  \item{gcContent}{GC-content of all sequence variants.}
  \item{tm}{Tm of all sequence variants.}
  \item{deltaG}{delta G (in kcal/mole) of all sequence variants.}
  \item{method}{Design method used to generate the oligo: "ambiguous",
  "mixedFwd" or "mixedRev".}
  \item{score}{Oligo score, the lower the better.
  See "Score" for more details.}
  \item{roiStart}{First position of the input \code{RprimerProfile} object
    (roi = region of interest).}
  \item{roiEnd}{Last position of the input \code{RprimerProfile} object.}
}
}

\section{Design strategy for primers}{


Primers can be generated by using one of the two following strategies:

\itemize{
\item{The \strong{ambiguous strategy} (default) generates primers from the
IUPAC consensus sequence, which means that degenerate bases can
occur at any position in the primer.
Probes are always designed using the ambiguous
strategy}
\item{The \strong{mixed strategy} generates primers from both the majority
and the IUPAC consensus sequence. These primers consist of a shorter
degenerate part at the 3' end (approx. 1/3 of the primer, targeting a
conserved
region) and a longer consensus part at the 5' end (approx.
2/3 of the primer),
which instead of having ambiguous bases contains the most probable
nucleotide at each position.
This strategy is inspired by the Consensus-Degenerate Hybrid
Oligonucleotide Primer (CODEHOP) principle (Rose et al., 1998).
The idea is that the degenerate 3' end part will bind specifically to
the target sequence in the initial PCR cycles, and promote amplification
in spite of eventual mismatches at the 5' consensus part
(since 5' end mismatches are generally less detrimental than
3' end mismatches). In this way, the generated products
will match the 5' ends of all primers perfectly, which allows them
to be efficiently amplified in later PCR cycles.
To provide a sufficiently high tm in spite of mismatches, it is
recommended to design relatively long primers (> 25 bases) when using
this strategy}
}
}

\section{Validity checks}{


For an oligo to be considered as valid, all sequence variants must fulfill
all the specified design constraints.

Oligos with sequence variants containing
more than four consecutive runs
of the same
nucleotide (e.g. "AAAAA") and/or more than three consecutive runs
of the same di-nucleotide (e.g. "TATATATA") are excluded from consideration.
}

\section{Tm and delta G}{


Melting temperature and delta G (change in Gibbs free energy)
are calculated for perfectly matching
DNA duplexes using the
nearest-neighbor
method (SantaLucia and Hicks, 2004).
Table values, salt correction method and equations are
from SantaLucia and Hicks, 2004.
The table values can also be
found by calling \code{rprimer:::lookup$nn}.
}

\section{Score}{


All valid oligos are scored based on their identity, coverage,
degeneracy,
average GC content and tm range. These scores are summarized,
and the weight of each individual score is 1. The lowest, and best,
possible score is 0, and the worst possible score is 12.

The scoring scheme is presented below.

\strong{Identity and coverage}

\tabular{lr}{
Value range \tab Score \cr
(0.99, 1] \tab 0 \cr
(0.95, 0.99] \tab 1 \cr
(0.90, 0.95] \tab 2 \cr
<= 0.90 \tab 3
}

\strong{Degeneracy}

\tabular{lr}{
Value \tab Score \cr
1 \tab 0 \cr
2 or 3 \tab 1 \cr
3 or 4 \tab 2 \cr
More than 4 \tab 3
}

\strong{Average GC-content}

The score for the average GC-content is based on how much
it deviates (in absolute value) from 0.5.

\tabular{lr}{
Value range \tab Score \cr
[0, 0.05)
\tab 0 \cr
[0.05, 0.1)
\tab 1 \cr
[0.1,  0.2)
\tab 2 \cr
>= 0.2 \tab 3
}

\strong{Tm range}

\tabular{lr}{
Value \tab Score \cr
[0, 1)
\tab 0 \cr
[1, 2)
\tab 1 \cr
[2, 3)
\tab 2 \cr
>= 3 \tab 3
}
}

\examples{
data("exampleRprimerProfile")
target <- exampleRprimerProfile

## Design primers and probes with default values
oligos(target)

## Design primers and probes only within a specific region of interest
roi <- target[target$position >= 5000 & target$position <= 6000, ]
oligos(roi)

## Design primers only
oligos(roi, probe = FALSE)

## Allow higher degeneracy
oligos(roi,
    maxDegeneracyPrimer = 32,
    probe = FALSE
)
}
\references{
Rose, Timothy M, Emily R Schultz, Jorja G Henikoff, Shmuel Pietrokovski,
Claire M McCallum, and Steven Henikoff. 1998. Consensus-Degenerate
Hybrid Oligonucleotide Primers for
Amplification of Distantly Related Sequences. Nucleic Acids Research 26 (7):
1628-35.

SantaLucia Jr, J., & Hicks, D. (2004).
The thermodynamics of DNA structural motifs.
Annu. Rev. Biophys. Biomol. Struct., 33, 415-440.
}
