% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotData}
\alias{plotData}
\alias{plotData,RprimerProfile-method}
\title{Plot an Rprimer-object (generic)}
\usage{
plotData(x, ...)

\S4method{plotData}{RprimerProfile}(x, shadeFrom = NULL, shadeTo = NULL)
}
\arguments{
\item{x}{An \code{RprimerProfile} object.}

\item{...}{Additional arguments that should be passed to the plot.}

\item{shadeFrom}{Optional. If a particular area should be shaded, where the shade
should start (an integer).}

\item{shadeTo}{Optional. If a particular area should be shaded, where the shade
should end (an integer).}
}
\value{
A plot.
}
\description{
Plot an Rprimer-object (generic)

Plot an RprimerProfile-object (method)
}
\section{Methods (by class)}{
\itemize{
\item \code{RprimerProfile}: 
}}

\examples{
data("exampleRprimerProfile")
plotData(exampleRprimerProfile)
plotData(exampleRprimerProfile, shadeFrom = 500, shadeTo = 1000)

}
