% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotData}
\alias{plotData}
\alias{plotData,RprimerProfile-method}
\alias{plotData,RprimerOligo-method}
\alias{plotData,RprimerAssay-method}
\title{Plot an Rprimer-object (generic)}
\usage{
plotData(x, ...)

\S4method{plotData}{RprimerProfile}(x, type = "overview", ...)

\S4method{plotData}{RprimerOligo}(x)

\S4method{plotData}{RprimerAssay}(x)
}
\arguments{
\item{x}{An \code{RprimerProfile}, \code{RprimerOligo} or \code{RprimerAssay} object.}

\item{...}{Optional arguments when plotting \code{RprimerProfile} objects:

For \code{type = "overview"}:

\code{highligt}: If a specific genomic region should be highlighted.
A numeric vector, e.g. \code{c(100, 1000)}, defaults to \code{NULL}.

For \code{type = "nucleotide"}:

\code{rc}: If the plotted sequence should be displayed
as reverse complement or not.
\code{TRUE} or {FALSE}, defaults to \code{FALSE}.}

\item{type}{For code{RprimerProfile}-objects.
Which type of plot that should be generated: "overview", or "nucleotide",
defaults to "overview".}
}
\value{
A plot.
}
\description{
Plot an Rprimer-object (generic)

Plot an RprimerProfile-object (method)

Plot an RprimerOligo-object (method)

Plot an RprimerAssay-object (method)
}
\section{Methods (by class)}{
\itemize{
\item \code{RprimerProfile}: 

\item \code{RprimerOligo}: 

\item \code{RprimerAssay}: 
}}

\examples{
## Plot an RprimerProfile object

data("exampleRprimerProfile")
prof <- exampleRprimerProfile

## Plot an overwiev
plotData(prof)

## Highlight a specific area
plotData(prof, highlight = c(500, 1000))

## Select a region of interest
roi <- prof[prof$position >= 500 & prof$position <= 550, ]

## Plot an overview of the roi
plotData(roi)

## Plot the nucleotide distribution of the roi
plotData(roi, type = "nucleotide")

## Plot the nucleotide distribution of the roi, as reverse complement
plotData(roi, type = "nucleotide", rc = TRUE)

## Plot an RprimerOligo object

data("exampleRprimerOligo")
oligos <- exampleRprimerOligo

## Plot all oligos
plotData(oligos)

## Select a subset of the oligos, and plot
selected <- oligos[oligos$start >= 5000 & oligos$end <= 5500, ]
plotData(selected)

## Plot an RrimerAssay object
data("exampleRprimerAssay")
plotData(exampleRprimerAssay)
}
