% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotNucleotides}
\alias{plotNucleotides}
\title{Plot nucleotide distribution}
\usage{
plotNucleotides(x, from = NULL, to = NULL, rc = FALSE)
}
\arguments{
\item{x}{An \code{RprimerProfile} object.}

\item{from}{Optional. Start position (an integer).}

\item{to}{Optional. End position (an integer).}

\item{rc}{\code{TRUE} or {FALSE}. If the plotted sequence should be displayed
as reverse complement or not. Defaults to \code{FALSE}.}
}
\value{
A barplot with nucleotide distribution.
}
\description{
Plot nucleotide distribution
}
\examples{
data("exampleRprimerProfile")
plotNucleotides(exampleRprimerProfile, from = 1, to = 10)
## Plot the first 10 bases
plotNucleotides(exampleRprimerProfile, from =  1, to = 10, rc = TRUE)
## As reverse complement

}
