% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNucleotides.R
\name{plotNucleotides}
\alias{plotNucleotides}
\title{Plot nucleotide distribution}
\usage{
plotNucleotides(x, rc = FALSE)
}
\arguments{
\item{x}{An \code{RprimerProfile} object.}

\item{rc}{If the plotted sequence should be displayed
as reverse complement or not.
\code{TRUE} or {FALSE}, defaults to \code{FALSE}.}
}
\value{
A barplot with nucleotide distribution.
}
\description{
Plot nucleotide distribution
}
\examples{
## Load example data
data("exampleRprimerProfile")
x <- exampleRprimerProfile

## Plot the first 10 bases
roi <- x[x$position >= 1 & x$position <= 10, ]
plotNucleotides(roi)

## As reverse complement
plotNucleotides(roi, rc = TRUE)
}
