% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta_alignment.R
\name{read_fasta_alignment}
\alias{read_fasta_alignment}
\title{Read an alignment in fasta format}
\usage{
read_fasta_alignment(x)
}
\arguments{
\item{x}{The name of the file of which the alignment is to be read from
(a character vector of length one).}
}
\value{
The alignment from the input
file (a named list with class attribute 'rprimer_alignment').
Each object (DNA sequence) has the same name as in the input file,
except for the '>' symbol.
Sequences are presented as character vectors of length one,
in lowercase format.
}
\description{
\code{read_fasta_alignment} reads a .txt-file with an alignment of
DNA sequences in fasta format.
}
\details{
The file must be a .txt file and contain one or more aligned
DNA sequences in fasta format. Sequences in fasta format always
begins with a '>' symbol, followed by a single-line name, with the
sequence present on the next line.
All sequences (including gaps) must be of the same length.
Valid nucleotides are 'a', 'c', 'g', 't', 'r', 'y', 'm', 'k', 's', 'w',
n', 'h', 'd', 'v', 'b' and '-'. The alignment can be in either
upper- or lowercase format.
}
\note{
This function is partly inspired by \code{seqinr::read.fasta}
(see references).
}
\examples{
read_fasta_alignment(
system.file('extdata', 'example_alignment.txt', package = 'rprimer')
)

}
\references{
Charif, D. and Lobry, J.R. (2007)
SeqinR 1.0-2: a contributed package to the R project for statistical
computing devoted to biological sequences retrieval and analysis.
Structural approaches to sequence evolution: Molecules, networks,
populations, pp 207-232.
}
