% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{read_fasta_alignment}
\alias{read_fasta_alignment}
\title{Read an alignment in fasta format}
\usage{
read_fasta_alignment(x)
}
\arguments{
\item{x}{The name of the file of which the
alignment is to be read from (a character vector of length one).}
}
\value{
The alignment from the input
file (a named list with class attribute 'rprimer_alignment').
Each object (DNA sequence) will have the same name as its
sequence name in the input file, except for the '>' symbol.
Sequences are will be presented as character vectors of length one,
in lowercase format.
}
\description{
\code{read_fasta_alignment} reads a fasta file with aligned DNA
sequences.
}
\details{
The file must contain one or more aligned
DNA sequences in fasta format. Sequences in fasta format
begins with a '>' symbol, followed by a single-line name.
The sequence is present on the next line.
All sequences (including gaps) must be of the same length,
and the minimum length of the alignment is 200.
Valid nucleotides are 'a', 'c', 'g', 't', 'r', 'y', 'm', 'k', 's', 'w',
n', 'h', 'd', 'v', 'b' and '-'. The alignment can be in either
upper- or lowercase format.
}
\note{
This function is based on ##########################################
}
\examples{
read_fasta_alignment(
system.file('extdata', 'example_alignment.txt', package = 'rprimer')
)

}
