% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{remove_gaps}
\alias{remove_gaps}
\title{Remove positions with high gap frequency}
\usage{
remove_gaps(x, threshold = 0.5)
}
\arguments{
\item{x}{An alignment of DNA sequences
(an object of class 'rprimer_alignment').}

\item{threshold}{A number between 0.5 and <1 (the default is 0.5).}
}
\value{
An alignment (an list of class 'rprimer_alignment'), where
all positions with a gap proportion higher than the stated threshold
are removed.
}
\description{
\code{remove_gaps} removes all positions with a gap frequency
higher than a stated threshold in an alignment of DNA sequences.
}
\details{
Gaps are recognised as "-".
Note that the positions will
not be kept from the input alignment. The positions in the new
alignment will always start at 1, and increase by 1 for every new
position.
}
\examples{
remove_gaps(example_rprimer_alignment, threshold = 0.5)

}
\seealso{
\code{read_fasta_alignment}
}
