% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_plot.R
\name{rp_plot}
\alias{rp_plot}
\alias{rp_plot.rprimer_alignment}
\alias{rp_plot.rprimer_profile}
\alias{rp_plot.rprimer_properties}
\title{Plot an rprimer object (generic)}
\usage{
rp_plot(x, ...)

\method{rp_plot}{rprimer_alignment}(x, ...)

\method{rp_plot}{rprimer_profile}(x, rc = FALSE, ...)

\method{rp_plot}{rprimer_properties}(x, ...)
}
\arguments{
\item{x}{An rprimer object (see methods for details).}

\item{...}{Additional arguments that should be passed to the plot, when
plotting an object of class 'rprimer_profile'.}

\item{rc}{\code{TRUE/FALSE}. If the plotted sequence should be displayed
as reverse complement or not. The default is \code{FALSE}.}
}
\value{
A plot.
}
\description{
Plot an rprimer object (generic)
}
\section{Methods (by class)}{
\itemize{
\item \code{rprimer_alignment}: Plot an rprimer_alignment object.

\item \code{rprimer_profile}: Plot an rprimer_profile object.

\item \code{rprimer_properties}: Plot an rprimer_properties object.
}}

\examples{
rp_plot(example_rprimer_alignment)
rp_plot(example_rprimer_profile[ , 1:20], rc = TRUE)
rp_plot(example_rprimer_properties)
rp_plot(example_rprimer_properties[1:20, ])

}
