% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{rp_plot}
\alias{rp_plot}
\alias{rp_plot.rprimer_alignment}
\alias{rp_plot.rprimer_sequence_profile}
\alias{rp_plot.rprimer_sequence_properties}
\title{Plot an rprimer-object (generic)}
\usage{
rp_plot(x, ...)

\method{rp_plot}{rprimer_alignment}(x, ...)

\method{rp_plot}{rprimer_sequence_profile}(x, ..., from = NULL, to = NULL, rc = FALSE)

\method{rp_plot}{rprimer_sequence_properties}(x, ...)
}
\arguments{
\item{x}{An rprimer-object (see methods for details).}

\item{from}{At which position the plot begins (an integer).}

\item{to}{At which position the plot ends (an integer).}

\item{rc}{\code{TRUE/FALSE}. If the plotted sequence should be displayed
as reverse complement or not. The default is \code{FALSE}.}
}
\value{
A plot.
}
\description{
Plot an rprimer-object (generic)
}
\section{Methods (by class)}{
\itemize{
\item \code{rprimer_alignment}: Plot an object of class 'rprimer_alignment'.

\item \code{rprimer_sequence_profile}: Plot an object of class 'rprimer_sequence_profile'.

\item \code{rprimer_sequence_properties}: Plot an object of class 'rprimer_sequence_properties'.
}}

\examples{
rp_plot(example_rprimer_alignment)
rp_plot(example_rprimer_sequence_profile, from = 1, to = 20, rc = TRUE)
rp_plot(example_rprimer_sequence_properties)

}
