% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_save.R
\name{rp_save}
\alias{rp_save}
\alias{rp_save.rprimer_alignment}
\alias{rp_save.rprimer_properties}
\alias{rp_save.rprimer_oligo}
\alias{rp_save.rprimer_assay}
\title{Save an rprimer object to a file (generic)}
\usage{
rp_save(x, filename)

\method{rp_save}{rprimer_alignment}(x, filename)

\method{rp_save}{rprimer_properties}(x, filename)

\method{rp_save}{rprimer_oligo}(x, filename)

\method{rp_save}{rprimer_assay}(x, filename)
}
\arguments{
\item{x}{An rprimer object (see methods for details).}

\item{filename}{The name of the file. A character vector of length one.}
}
\value{
A saved file.
}
\description{
Save an rprimer object to a file (generic)
}
\details{
The file format will depend on object.
Objects of class rprimer_alignment will be saved as .txt-files,
whereas objects of class rprimer_properties, rprimer_oligo
and rprimer_assay will be saved as .csv-files.
}
\section{Methods (by class)}{
\itemize{
\item \code{rprimer_alignment}: Save an rprimer_alignment object.

\item \code{rprimer_properties}: Save an rprimer_properties object.

\item \code{rprimer_oligo}: Save an rprimer_oligo object.

\item \code{rprimer_assay}: Save an'rprimer_assay' object.
}}

\examples{
\dontrun{
  rp_save(example_rprimer_alignment, 'my_alignment')
  rp_save(example_rprimer_properties, 'my_sequence_properties')
  rp_save(example_rprimer_oligo, 'my_oligos')
  rp_save(example_rprimer_assay, 'my_assays')
}

}
