% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprimer-package.R
\docType{package}
\name{rprimer-package}
\alias{rprimer}
\alias{rprimer-package}
\title{rprimer: Design (RT)-(q/d)PCR Assays from a Multiple DNA Sequence Alignment}
\description{
Functions and workflow to design
    primers, probes and (RT)-(q/d)PCR assays from a multiple DNA sequence 
    alignment. The design process is especially developed
    for sequence variable targets, such as RNA viruses. 
    Please see the package vignette for more information.
}
\details{
The package contains four functions: \code{consensusProfile()},
\code{oligos()}, \code{assays()} and \code{plotData()}.

For more information about rprimer and its
functionality, please see the package vignette.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sofpn/rprimer}
  \item Report bugs at \url{https://github.com/sofpn/rprimer/issues}
}

}
\author{
\strong{Maintainer}: Sofia Persson \email{sofiapersson27@gmail.com} (\href{https://orcid.org/0000-0003-2611-3030}{ORCID})

}
\keyword{internal}
