% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprimer-package.R
\docType{package}
\name{rprimer-package}
\alias{rprimer}
\alias{rprimer-package}
\title{rprimer: Design (RT)-(q/d)PCR Assays from a Multiple DNA Sequence Alignment}
\description{
Functions and workflow for visualizing sequence conservation 
    and designing
    degenerate primers, probes and (RT)-(q/d)PCR assays from a 
    multiple DNA sequence alignment. The workflow is especially developed for
    sequence variable viruses.
}
\details{
The package contains five functions: \code{consensusProfile()},
\code{oligos()}, \code{assays()}, \code{plotData()} and
\code{convertToDNAStringSet()}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sofpn/rprimer}
  \item Report bugs at \url{https://github.com/sofpn/rprimer/issues}
}

}
\author{
\strong{Maintainer}: Sofia Persson \email{sofiapersson27@gmail.com} (\href{https://orcid.org/0000-0003-2611-3030}{ORCID})

}
\keyword{internal}
