% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{RprimerMatchAssay}
\alias{RprimerMatchAssay}
\title{RprimerMatchAssay}
\usage{
RprimerMatchAssay(...)
}
\arguments{
\item{...}{A data frame or list to be converted into an \code{RprimerMatchAssay}
object.}
}
\value{
An \code{RprimerMatchAssay} object if validation succeeds, an error
message otherwise.
}
\description{
The constructor, \code{RprimerMatchAssay()},
constructs an \code{RprimerMatchAssay} object in a similar fashion as the
\code{S4Vectors::DataFrame()} constructor.
}
\section{Functions}{
\itemize{
\item \code{RprimerMatchAssay}: 
}}

\examples{
data("exampleRprimerMatchAssay")
x <- as.data.frame(exampleRprimerMatchAssay)
RprimerMatchAssay(x)
}
\references{
Pages, H., Lawrence, M., and Aboyoun, R. (2020). S4Vectors:
Foundation of vector-like and list-like containers in
Bioconductor. R package version 0.28.0.
}
