% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RprimerOligo-class}
\alias{RprimerOligo-class}
\alias{.RprimerOligo}
\alias{RprimerOligo}
\title{An S4 class for representation of oligos}
\usage{
RprimerOligo(...)
}
\arguments{
\item{...}{A data frame or list to be converted into an \code{RprimerOligo} object.}
}
\value{
An \code{RprimerOligo} object if validation succeeds, an error
message otherwise.
}
\description{
\code{RprimerOligo} extends the \code{S4Vectors::DataFrame} class,
without any additional slots.

The constructor, \code{RprimerOligo()},
constructs an \code{RprimerOligo} object in a similar fashion as the
\code{S4Vectors::DataFrame()} constructor.
}
\section{Functions}{
\itemize{
\item \code{RprimerOligo}: 
}}

\examples{
data("exampleRprimerOligo")
x <- as.data.frame(exampleRprimerOligo)
RprimerOligo(x)
}
\references{
Pages, H., Lawrence, M., and Aboyoun, R. (2020). S4Vectors:
Foundation of vector-like and list-like containers in
Bioconductor. R package version 0.28.0.
}
\seealso{
Other RprimerOligo: 
\code{\link{coerce}()}
}
\concept{RprimerOligo}
