% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprimerProfile.R
\docType{class}
\name{RprimerProfile-class}
\alias{RprimerProfile-class}
\alias{RprimerProfile}
\title{An S4 class for representation of an alignment profile}
\usage{
RprimerProfile(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}
}
\value{
An RprimerProfile object if validation succeeds. An error
message if not.
}
\description{
An extension of the S4 class SummarizedExperiment::SummarizedExperiment.
}
\details{
A numeric matrix with the proportion of each
nucleotide at each position within an alignment
of DNA sequences. The matrix has six rows,
named 'A', 'C', 'G', 'T', '-' and 'Other'. '-' represents gaps and
'Other' represents nucleotides other than A, C, G and T, e.g.
wobble bases. The columns are named
according to which position they correspond to in the alignment.
}
\section{Functions}{
\itemize{
\item \code{RprimerProfile}: 
}}

