% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RprimerProfile-class}
\alias{RprimerProfile-class}
\alias{.RprimerProfile}
\title{An S4 class for representation of a consensus profile}
\description{
\code{RprimerProfile} extends the \code{S4Vectors::DataFrame} class,
without any additional slots. It has the same accessors and
coercion, subsetting, and combining methods as the parent class, but has
some additional checks for validity.
}
