% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprimerProfile.R
\docType{class}
\name{RprimerProfile}
\alias{RprimerProfile}
\title{An S4 class for representation of an alignment profile}
\description{
An S4 class for representation of an alignment profile
}
\section{Slots}{

\describe{
\item{\code{assay}}{A numeric matrix with the proportion of each
nucleotide at each position within an alignment
of DNA sequences. The matrix has six rows,
named 'A', 'C', 'G', 'T', '-' and 'Other'. '-' represents gaps and
'Other' represents nucleotides other than A, C, G and T, e.g.
wobble bases. The columns are named
according to which position they correspond to in the alignment.}
}}

