% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{RprimerProfile}
\alias{RprimerProfile}
\title{RprimerProfile}
\usage{
RprimerProfile(...)
}
\arguments{
\item{...}{A data frame or list to be converted into an \code{RprimerProfile} object.}
}
\value{
An \code{RprimerAssay} object if validation succeeds, an error
message otherwise.
}
\description{
Constructs an \code{RprimerProfile} object in a similar fashion as the
\code{S4Vectors::DataFrame} constructor.
}
\examples{
data("exampleRprimerProfile")
x <- as.data.frame(exampleRprimerProfile)
RprimerProfile(x)
}
