% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRprimerApp.R
\name{runRprimerApp}
\alias{runRprimerApp}
\title{Run the rprimer package as a shiny application}
\usage{
runRprimerApp()
}
\value{
A Shiny web application
}
\description{
Run the rprimer package as a shiny application
}
\examples{
\donttest{runRprimerApp()}
}
