% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_rprimer_alignment.R
\name{[.rprimer_alignment}
\alias{[.rprimer_alignment}
\title{Extract elements in an rprimer_alignment object}
\usage{
\method{[}{rprimer_alignment}(x, i, ...)
}
\description{
Extract elements in an rprimer_alignment object
}
