% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_rprimer_profile.R
\name{[.rprimer_profile}
\alias{[.rprimer_profile}
\title{Extract elements in an rprimer_profile object}
\usage{
\method{[}{rprimer_profile}(x, i, ...)
}
\description{
Extract elements in an rprimer_profile object
}
