% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_rprimer_alignment.R
\name{[<-.rprimer_alignment}
\alias{[<-.rprimer_alignment}
\title{Replace elements in an rprimer_alignment object}
\usage{
\method{[}{rprimer_alignment}(x, i) <- value
}
\description{
Replace elements in an rprimer_alignment object
}
