% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{write_report}
\alias{write_report}
\title{Write assay report}
\usage{
write_report(
  filename = "my_assay_report",
  assay_selection,
  sequence_profile,
  sequence_properties
)
}
\arguments{
\item{filename}{(a character vector of length one)}

\item{assay_selection}{}

\item{sequence_profile}{}

\item{sequence_properties}{}
}
\value{
A html-report
}
\description{
Write assay report
}
\details{
Details:
rmarkdown and xxx are needed for this function.
}
