% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_report.R
\name{write_report}
\alias{write_report}
\title{Write assay report}
\usage{
write_report(
  assay_selection,
  sequence_profile,
  sequence_properties,
  comment = NULL,
  filename = "my_assay_report"
)
}
\arguments{
\item{assay_selection}{An rprimer_assay object, with one row.}

\item{sequence_profile}{An rprimer_profile object.}

\item{sequence_properties}{An rprimer_properties object.}

\item{comment}{Optional. Comments that should be included in the report.
A character vector of length one. The default is \code{NULL}.}

\item{filename}{The name of the file. A character vector of length one.}
}
\value{
A html-document with detailed assay information.
}
\description{
Write assay report
}
\details{
Details:
rmarkdown and kableExtra are needed for this function to work.
}
\examples{
\dontrun{
write_report(
example_rprimer_assay[1, ],
example_rprimer_profile,
example_rprimer_properties,
comment = "My specific comment",
filename = "my_report",
)
}

}
