% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_report.R
\name{write_report}
\alias{write_report}
\title{Write assay report}
\usage{
write_report(
  filename = "my_assay_report",
  assay_selection,
  sequence_profile,
  sequence_properties,
  comment = NULL
)
}
\arguments{
\item{filename}{(a character vector of length one).}

\item{assay_selection}{An object of class 'rprimer_assay', with one row.}

\item{sequence_profile}{An object of class 'rprimer_profile'.}

\item{sequence_properties}{An object of class 'rprimer_properties.}

\item{comment}{Optional. Comments that should be included in the report.
A character vector of length one. The default is \code{NULL}.}
}
\value{
A html-document with detailed assay information.
}
\description{
Write assay report
}
\details{
Details:
rmarkdown and kableExtra are needed for this function to work.
}
\examples{
\dontrun{
write_report("my_report",
example_rprimer_assay[1, ],
example_rprimer_profile,
example_rprimer_properties,
comment = "My specific comment"
)
}

}
